### 2023-03-16 #5262 @editorialbot
[PRE REVIEW]: chombo-discharge: An AMR code for gas discharge simulations in complex geometries
**Submitting author:** @rmrsk (<a href="http://orcid.org/0000-0003-1706-9736">Robert Marskar</a>)
**Repository:** https://github.com/chombo-discharge/chombo-discharge
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.0
