### 2023-03-17 #5262 @rmrsk
Sure @ppxasjsm  - from the the reviewer list it seems like @sigvaldm and @tjford could do it? I've also put out feelers to folks outside of the JOSS list of reviewers.
