### 2023-03-17 #5262 @kyleniemeyer
> Thanks @kyleniemeyer and @ppxasjsm!
>
> The DOIs point to the correct place. I've started fishing for reviewers, will suggest some names soon...

@rmrsk sorry for not being more clear—I meant specifically take a look at whether the missing DOIs that `editorialbot` reported should be added:

```
MISSING DOIs

- 10.1016/j.jcp.2019.03.036 may be a valid DOI for title: An adaptive Cartesian embedded boundary approach for fluid simulations of two- and three-dimensional low temperature plasma filaments in complex geometries
- 10.1088/1361-6463/aa8faf may be a valid DOI for title: Simulating streamer discharges in 3D with the parallel adaptive Afivo framework
- 10.1088/1361-6595/aca0be may be a valid DOI for title: Evolution of positive streamers in air over non-planar dielectrics: experiments and simulations
```
