+++ paper/paper.bib
@@ -2,6 +2,7 @@
    Author = {{Marskar}, R},
    Journal = {Journal of Computational Physics},
    Year = 2019,
+   doi = {10.1016/j.jcp.2019.03.036},
    url = {https://doi.org/10.1016/j.jcp.2019.03.036},
    Title = {An adaptive Cartesian embedded boundary approach for fluid simulations of two- and three-dimensional low temperature plasma filaments in complex geometries},
 }
@@ -42,6 +43,7 @@ abstract = {A 3D fluid model has been developed to simulate streamer discharges 
   doi={10.1109/TPS.2021.3120507}}
 @article{Teunissen:2017,
    Author = {{Teunissen}, J and {Ebert}, E},
+   doi = {10.1088/1361-6463/aa8faf},
    url = {https://doi.org/10.1088/1361-6463/aa8faf},
    Journal = {Journal of Physics D: Applied Physics},
    Month = {11},
@@ -52,6 +54,7 @@ abstract = {A 3D fluid model has been developed to simulate streamer discharges 
 }
 @article{Meyer:2022,
    Author = {{Meyer}, H K H and {Marskar}, R and {Mauseth}, F},
+   doi = {10.1088/1361-6595/aca0be},
    url = {https://doi.org/10.1088/1361-6595/aca0be},
    Journal = {Plasma Sources Science and Technology},
    Pages = {114006},
