### 2023-03-20 #5262 @rmrsk
@kyleniemeyer Ooops! They sure should - I was confused because the URLs used the DOI and indeed pointed to the correct place in the compiled paper. But yes, there was a missing doi = {...} in the bib file. Added them just now.
