### 2023-03-29 57f7106
Add transparent EB option for McPhoto
