### 2023-03-29 4b425d8
Input options inplace
