### 2023-04-01 89e709a
Added mask to mfhelmholtzopfactory
