### 2023-04-01 d9fbffe
Passed valid cells to factories
