+++ Source/RadiativeTransfer/CD_EddingtonSP1.cpp
@@ -880,6 +880,7 @@ EddingtonSP1::setupHelmholtzFactory()
                                                                                       m_alpha,
                                                                                       m_beta,
                                                                                       m_amr->getProbLo(),
+                                                                                      m_amr->getValidCells(m_realm),
                                                                                       levelGrids,
                                                                                       interpolator,
                                                                                       fluxReg,
