+++ Source/Elliptic/CD_EBHelmholtzOp.cpp
@@ -141,7 +141,7 @@ EBHelmholtzOp::EBHelmholtzOp(const Location::Cell                             a_
   // Define BC objects.
   const int ghostCF = m_hasCoar ? m_interpolator->getGhostCF() : 99;
   m_domainBc->define(m_dataLocation, m_eblg, m_probLo, m_dx);
-  m_ebBc->define(m_dataLocation, m_eblg, m_probLo, m_dx, ghostCF);
+  m_ebBc->define(m_dataLocation, m_eblg, m_eblgFine, m_probLo, m_dx, m_hasFine, m_refToFine, ghostCF);
 
   // Define stencils and compute relaxation terms.
   this->defineStencils();
