+++ Source/Elliptic/CD_EBHelmholtzEBBC.cpp
@@ -27,8 +27,11 @@ EBHelmholtzEBBC::~EBHelmholtzEBBC() { CH_TIME("EBHelmholtzEBBC::~EBHelmholtzEBBC
 void
 EBHelmholtzEBBC::define(const Location::Cell a_dataLocation,
                         const EBLevelGrid&   a_eblg,
+                        const EBLevelGrid&   a_eblgFine,
                         const RealVect&      a_probLo,
                         const Real&          a_dx,
+                        const bool           a_hasFineAMRLevel,
+                        const int            a_refToFine,
                         const int            a_ghostCF)
 {
   CH_TIME(
@@ -37,11 +40,14 @@ EBHelmholtzEBBC::define(const Location::Cell a_dataLocation,
   CH_assert(a_dx > 0.0);
   CH_assert(a_ghostCF >= 0);
 
-  m_dataLocation = a_dataLocation;
-  m_eblg         = a_eblg;
-  m_probLo       = a_probLo;
-  m_dx           = a_dx;
-  m_ghostCF      = a_ghostCF;
+  m_dataLocation    = a_dataLocation;
+  m_eblg            = a_eblg;
+  m_eblgFine        = a_eblgFine;
+  m_probLo          = a_probLo;
+  m_dx              = a_dx;
+  m_hasFineAMRLevel = a_hasFineAMRLevel;
+  m_refToFine       = a_refToFine;
+  m_ghostCF         = a_ghostCF;
 
   this->define();
 }
