+++ Source/Elliptic/CD_EBHelmholtzOpFactory.cpp
@@ -24,6 +24,7 @@ EBHelmholtzOpFactory::EBHelmholtzOpFactory(const Location::Cell    a_dataLocatio
                                            const Real&             a_alpha,
                                            const Real&             a_beta,
                                            const RealVect&         a_probLo,
+                                           const AMRMask&          a_amrValidCells,
                                            const AmrLevelGrids&    a_amrLevelGrids,
                                            const AmrInterpolators& a_amrInterpolators,
                                            const AmrFluxRegisters& a_amrFluxRegisters,
@@ -51,6 +52,7 @@ EBHelmholtzOpFactory::EBHelmholtzOpFactory(const Location::Cell    a_dataLocatio
 
   m_probLo = a_probLo;
 
+  m_amrValidCells    = a_amrValidCells;
   m_amrLevelGrids    = a_amrLevelGrids;
   m_amrInterpolators = a_amrInterpolators;
   m_amrFluxRegisters = a_amrFluxRegisters;
