+++ Source/Elliptic/CD_MFHelmholtzOpFactory.cpp
@@ -654,6 +654,7 @@ MFHelmholtzOpFactory::MGnewOp(const ProblemDomain& a_fineDomain, int a_depth, bo
                              MFLevelGrid(),
                              MFLevelGrid(),
                              mflgMgCoar,
+                             RefCountedPtr<LevelData<BaseFab<bool>>>(),
                              interpolator,
                              fluxReg,
                              coarsener,
@@ -736,6 +737,7 @@ MFHelmholtzOpFactory::AMRnewOp(const ProblemDomain& a_domain)
                                         mflgCoFi,
                                         mflgCoar,
                                         mflgCoarMG,
+                                        m_amrValidCells[amrLevel],
                                         m_amrInterpolators[amrLevel],
                                         m_amrFluxRegisters[amrLevel],
                                         m_amrCoarseners[amrLevel],
