### 2023-04-01 1538ecc
Added ghost cells to valid cell mask.
