+++ Source/Elliptic/CD_MFHelmholtzOpFactory.cpp
@@ -27,32 +27,33 @@ constexpr int MFHelmholtzOpFactory::m_comp;
 constexpr int MFHelmholtzOpFactory::m_nComp;
 constexpr int MFHelmholtzOpFactory::m_mainPhase;
 
-MFHelmholtzOpFactory::MFHelmholtzOpFactory(const MFIS&             a_mfis,
-                                           const Location::Cell    a_dataLocation,
-                                           const Real&             a_alpha,
-                                           const Real&             a_beta,
-                                           const RealVect&         a_probLo,
-                                           const AMRMask&          a_amrValidCells,
-                                           const AmrLevelGrids&    a_amrLevelGrids,
-                                           const AmrInterpolators& a_amrInterpolators,
-                                           const AmrFluxRegisters& a_amrFluxRegisters,
-                                           const AmrCoarseners&    a_amrCoarseners,
-                                           const AmrRefRatios&     a_amrRefRatios,
-                                           const AmrResolutions&   a_amrResolutions,
-                                           const AmrCellData&      a_amrAcoef,
-                                           const AmrFluxData&      a_amrBcoef,
-                                           const AmrIrreData&      a_amrBcoefIrreg,
-                                           const DomainBCFactory&  a_domainBcFactory,
-                                           const EBBCFactory&      a_ebBcFactory,
-                                           const JumpBCFactory&    a_jumpBcFactory,
-                                           const IntVect&          a_ghostPhi,
-                                           const IntVect&          a_ghostRhs,
-                                           const Smoother&         a_smoother,
-                                           const ProblemDomain&    a_bottomDomain,
-                                           const int&              a_jumpOrder,
-                                           const int&              a_jumpWeight,
-                                           const int&              a_blockingFactor,
-                                           const AmrLevelGrids&    a_deeperLevelGrids)
+MFHelmholtzOpFactory::MFHelmholtzOpFactory(const MFIS&                               a_mfis,
+                                           const Location::Cell                      a_dataLocation,
+                                           const Real&                               a_alpha,
+                                           const Real&                               a_beta,
+                                           const RealVect&                           a_probLo,
+                                           const AMRMask&                            a_amrValidCells,
+                                           const AmrLevelGrids&                      a_amrLevelGrids,
+                                           const Vector<RefCountedPtr<MFLevelGrid>>& a_amrLevelGridsFiCo,
+                                           const AmrInterpolators&                   a_amrInterpolators,
+                                           const AmrFluxRegisters&                   a_amrFluxRegisters,
+                                           const AmrCoarseners&                      a_amrCoarseners,
+                                           const AmrRefRatios&                       a_amrRefRatios,
+                                           const AmrResolutions&                     a_amrResolutions,
+                                           const AmrCellData&                        a_amrAcoef,
+                                           const AmrFluxData&                        a_amrBcoef,
+                                           const AmrIrreData&                        a_amrBcoefIrreg,
+                                           const DomainBCFactory&                    a_domainBcFactory,
+                                           const EBBCFactory&                        a_ebBcFactory,
[…]
