### 2023-04-02 9dceb2a
Add warning in EBHelmholtzDirichletEBBC - will use eddington example as test case
