### 2023-04-02 c1d3a27
Add debug case for testing new EBHelmholtz functionality
