+++ Source/Elliptic/CD_EBHelmholtzDirichletEBBC.cpp
@@ -141,7 +141,7 @@ EBHelmholtzDirichletEBBC::define()
       // Try quadrants first.
       order = m_order;
       while (!foundStencil && order > 0) {
-        foundStencil = this->getLeastSquaresStencil(pairSten, vof, VofUtils::Neighborhood::Quadrant, dit(), order);
+        foundStencil = this->getGradStencilSingleLevel(pairSten, vof, VofUtils::Neighborhood::Quadrant, dit(), order);
         order--;
 
         // Check if stencil reaches too far across CF
@@ -153,7 +153,7 @@ EBHelmholtzDirichletEBBC::define()
       // If we couldn't find in a quadrant, try a larger neighborhood
       order = m_order;
       while (!foundStencil && order > 0) {
-        foundStencil = this->getLeastSquaresStencil(pairSten, vof, VofUtils::Neighborhood::Radius, dit(), order);
+        foundStencil = this->getGradStencilSingleLevel(pairSten, vof, VofUtils::Neighborhood::Radius, dit(), order);
         order--;
 
         // Check if stencil reaches too far across CF
@@ -207,9 +207,9 @@ EBHelmholtzDirichletEBBC::defineResidualStencils() noexcept
       BaseIVFAB<Real>&             gradPhiResidWeight       = m_boundaryWeightsResid[dit()];
       BaseIVFAB<VoFStencil>&       gradPhiResidStencils     = m_gradPhiResidStencils[dit()];
       BaseIVFAB<VoFStencil>&       gradPhiResidStencilsFine = m_gradPhiResidStencilsFine[dit()];
-      const BaseIVFAB<Real>&       gradPhiRelaxWeight     = m_boundaryWeightsRelax[dit()];
-      const BaseIVFAB<VoFStencil>& gradPhiRelaxStencil    = m_gradPhiRelaxStencils[dit()];
-      const BaseFab<bool>&         validCells             = (*m_amrValidCells)[dit()];
+      const BaseIVFAB<Real>&       gradPhiRelaxWeight       = m_boundaryWeightsRelax[dit()];
+      const BaseIVFAB<VoFStencil>& gradPhiRelaxStencil      = m_gradPhiRelaxStencils[dit()];
+      const BaseFab<bool>&         validCells               = (*m_amrValidCells)[dit()];
 
       // Where to define our residual stencils.
       const IntVectSet& ivs     = gradPhiRelaxStencil.getIVS();
@@ -223,8 +223,6 @@ EBHelmholtzDirichletEBBC::defineResidualStencils() noexcept
       // recompute a better stencil for the AMR residual.
       auto kernel = [&](const VolIndex& vof) -> void {
         const IntVect iv = vof.gridIndex();
-
-        bool useRelaxStencil = true;
 
         Vector<VolIndex> validVoFs;
         Vector<VolIndex> invalidVoFs;
@@ -248,12 +246,40 @@ EBHelmholtzDirichletEBBC::defineResidualStencils() noexcept
         if (invalidVoFs.size() > 0) {
           // We need to compute a new stencil for the finite volume flux, which also uses cells on
           // the fine level.
-#if 1 // This is just placeholder code to make things run while we adjust stencils
-          gradPhiResidWeight(vof, m_comp)   = gradPhiRelaxWeight(vof, m_comp);
-          gradPhiResidStencils(vof, m_comp) = gradPhiRelaxStencil(vof, m_comp);
 
-          std::cout << "need new stencil for vof = " << vof << std::endl;
-#endif
+          // Recompute a two-level stencil, replacing the invalid vofs with corresponding fine-level vofs
+          Vector<VolIndex> fineVoFs;
+          for (int i = 0; i < invalidVoFs.size(); i++) {
[…]

[…]
