### 2023-04-03 11a216b
Some warnings for myself so I know where to pick up next
