+++ Source/Elliptic/CD_EBHelmholtzEBBC.cpp
@@ -36,8 +36,10 @@ EBHelmholtzEBBC::define(const Location::Cell                           a_dataLoc
                         const int                                      a_refToFine,
                         const int                                      a_ghostCF)
 {
-  CH_TIME(
-    "EBHelmholtzEBBC::define(Location::Cell, EBLevelGrid, RefCountedPtr<LD<BaseIVFAB<Real> > >, RealVect, Real, int)");
+  CH_TIME("EBHelmholtzEBBC::define");
+
+  MayDay::Warning("EBHelmholtzEBBC::define -- need routine for computing the EB flux through the EB now!");
+  MayDay::Warning("EBHelmholtzEBBC::define -- check if AMRMultiGrid can sync so we don't interpolate twice");
 
   CH_assert(a_dx > 0.0);
   CH_assert(a_ghostCF >= 0);
