### 2023-04-04 #5335 @editorialbot
[REVIEW]: chombo-discharge: An AMR code for gas discharge simulations in complex geometries
**Submitting author:** @rmrsk (<a href="http://orcid.org/0000-0003-1706-9736">Robert Marskar</a>)
**Repository:** https://github.com/chombo-discharge/chombo-discharge
**Branch with paper.md** (empty if default branch): joss
**Version:** <v23.05>
