### 2023-04-11 41acde1
More simplifications to CI
