+++ .github/workflows/CI.yml
@@ -10,26 +10,26 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  # Formatting:
-  #   name: Format source
-  #   runs-on: ubuntu-latest
-  #   timeout-minutes: 10
+  Formatting:
+    name: Format source
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
 
-  #   strategy:
-  #     fail-fast: false
-  #     matrix:
-  #       directory: [Source, Physics, Geometries, Exec]
+    strategy:
+      fail-fast: false
+      matrix:
+        directory: [Source, Physics, Geometries, Exec]
 
-  #   steps:
-  #     - uses: actions/checkout@v3
-  #     - name: Run clang-format
-  #       uses: jidicula/clang-format-action@v4.10.1
-  #       with:
-  #         clang-format-version: '10'
-  #         check-path: ${{ matrix.directory }}
+    steps:
+      - uses: actions/checkout@v3
+      - name: Run clang-format
+        uses: jidicula/clang-format-action@v4.10.1
+        with:
+          clang-format-version: '10'
+          check-path: ${{ matrix.directory }}
 
   Linux-GNU:
-#    needs: Formatting
+    needs: Formatting
     name: Linux-GNU
     runs-on: ubuntu-latest
     timeout-minutes: 90
@@ -62,9 +62,9 @@ jobs:
         run: |
           cp $DISCHARGE_HOME/Lib/Local/GitHub/Make.defs.linux-gcc $CHOMBO_HOME/mk/Make.defs.local
 
-      - name: Build Chombo
-        run: |
-          make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} USE_HDF=${{ matrix.hdf }} MPI=${{ matrix.mpi }} chombo
+      # - name: Build Chombo
+      #   run: |
+      #     make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} USE_HDF=${{ matrix.hdf }} MPI=${{ matrix.mpi }} chombo
 
       # - name: Build chombo-discharge/source
       #   run: |
@@ -78,9 +78,50 @@ jobs:
       #   run: |
       #     make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} USE_HDF=${{ matrix.hdf }} MPI=${{ matrix.mpi }} discharge-physics          
 
-      
+  Build-documentation:
+    needs: Formatting
+    runs-on : ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Install depdencies
+        run: |
+          sudo apt-get update
+          sudo apt install doxygen
+          sudo apt install graphviz
+          sudo apt install texlive
+          sudo apt install texlive-latex-extra
+          sudo apt install latexmk
+          python3 -m pip install sphinx==5.0.0 sphinx_rtd_theme sphinxcontrib-bibtex
+
+      - name: Build doxygen
+        run: |
+          doxygen Docs/doxygen.conf
+
+      - name: Build HTML documentation
+        working-directory: ${{ github.workspace }}/Docs/Sphinx        
+        run: |
+          make html
+
+      - name: Build PDF documentation
+        working-directory: ${{ github.workspace }}/Docs/Sphinx
+        run: |
+          make latexpdf
+
+      - name: Upload
+        uses: actions/upload-artifact@v1
+        with:
+          name: docs
+          path: Docs/Sphinx/build
+
+      - name: Cleanup
+        working-directory: ${{ github.workspace }}/Docs
+        run: |
+          cp -a Sphinx/build/html/* ./
+          mv Sphinx/build/latex/chombo-discharge.pdf ./
+          rm -rf Sphinx/build          
[…]
