+++ Exec/Tests/KineticMonteCarlo/Avalanche/program.cpp
@@ -87,7 +87,7 @@ main(int argc, char* argv[])
     else if (alg == "tau") {
       nextDt = stopTime / numSteps;
 
-      kmcSolver.advanceTau(state, nextDt);
+      kmcSolver.advanceTauPlain(state, nextDt);
     }
     else if (alg == "hybrid") {
       nextDt = stopTime / numSteps;
