+++ Exec/Tests/README.md
@@ -15,13 +15,7 @@ python3 tests.py -suite AdvectionDiffusion
 
 ## Other options
 Use various flags for controlling how the applications are run and whether or not benchmark files for comparison will be generated.
-E.g.,
 
-```shell
-python3 tests.py --compile --clean --silent --no_compare
-```
-
-will do a clean compile of all tests and generate benchmark files.
 Configrable options are
 
 * ```--compile``` Force application to compile first.
@@ -29,15 +23,23 @@ Configrable options are
 * ```--clean``` Do a clean compilation.
 * ```--benchmark``` Generate benchmark files.
 * ```--no_exec``` Compile, but do not run tests.
-* ```--no_compare``` Run, but do not compare with benchmark files.
-* ```--parallel``` Compile with MPI.
-* ```-cores``` Run with specifies number of cores.
+* ```--compare``` Run, and then compare with benchmark files.
+* ```-dim```Test dimensionality. E.g. ```-dim 2``` does all 2d tests. If ```-dim``` is not 2 or 3, both 2d and 3d are run. 
+* ```-mpi``` (true/false) Use MPI or not
+* ```-hdf``` (true/false) Use HDF5 or not
+* ```-cores``` Run with specified number of cores.
 * ```-suites``` Tests suites to run. E.g. ```-suite AdvectionDiffusion Electrostatics```.
 * ```-tests``` Specific tests to run. E.g. ```-tests AdvectionDiffusion/Godunov```.
-* ```-dim```Test dimensionality. E.g. ```-dim 2``` does all 2d tests. If ```-dim``` is not 2 or 3, both 2d and 3d are run. 
 
-## Comparing benchmark files
-To compare the benchmark files with a new run, do
+E.g.,
+
 ```shell
-python3 tests.py --compile --clean --silent
+python3 tests.py --compile --clean --silent --benchmark -dim=2 -mpi=true -hdf=true -cores 12
+```
+
+will do a clean compile of all tests and generate benchmark files (using MPI and HDF5).
+If one wants to compare previous output with benchmark files, one may later do
+
+```shell
+python3 tests.py --compile --silent --compare -dim=2 -mpi=true -hdf=true -cores 12
 ```
