+++ .github/workflows/CI.yml
@@ -82,14 +82,14 @@ jobs:
         working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
         if: always()        
         run: |
-          python3 tests.py --compile --clean --no_exec -cores ${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }} -hdf=$ {{ matrix.hdf }}
+          python3 tests.py --compile --clean --no_exec -cores=${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }} -hdf=$ {{ matrix.hdf }}
           if [[ $? == 0 ]]; then exit 0; else exit 1; fi
 
       - name: Run tests
         working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
         if: always()        
         run: |
-          python3 tests.py-cores ${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }}
+          python3 tests.py -cores=${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }}
           if [[ $? == 0 ]]; then exit 0; else exit 1; fi                                        
 
   Build-documentation:
