+++ .github/workflows/linux-intel.yml
@@ -1,166 +1,0 @@
-name: Linux Intel
-
-on:
-  pull_request:
-    types: [push, opened, synchronize, reopened, ready_for_review]
-
-concurrency:
-  group: ${{ github.ref }}-${{ github.head_ref }}-linux-intel  
-  cancel-in-progress: true
-
-jobs:
-  serial-no-hdf:
-    if: github.event.pull_request.draft == false  
-    runs-on: ubuntu-latest
-    timeout-minutes: 300
-    strategy:
-      fail-fast: false
-      matrix:
-        dim: [2, 3]
-    name: Intel serial no-HDF5 DIM=${{ matrix.dim }}    
-    
-    env: 
-      DISCHARGE_HOME: ${{github.workspace}}
-      CHOMBO_HOME: ${{github.workspace}}/Submodules/Chombo-3.3/lib
-      CORES: 2            
-
-    steps:
-      - name: Checkout chombo-discharge
-        uses: actions/checkout@v2
-        with:
-          submodules: true
-
-      - name: Install dependencies
-        run: |
-          sudo apt-get update
-          sudo apt-get -y --no-install-recommends install csh libgetopt-complete-perl
-
-      - name: Install Intel compilers/MPI
-        working-directory: /tmp        
-        run: |
-          wget https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
-          sudo apt-key add GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
-          sudo add-apt-repository "deb https://apt.repos.intel.com/oneapi all main"
-          sudo apt-get -y install intel-oneapi-compiler-dpcpp-cpp-and-cpp-classic intel-oneapi-compiler-fortran intel-oneapi-mkl intel-oneapi-mpi intel-oneapi-mpi-devel intel-oneapi-openmp
-      
-      - name: Setup compilers
-        run: |
-          source /opt/intel/oneapi/setvars.sh
-        
-      - name: Copy makefile
-        run: |
-          cp $DISCHARGE_HOME/Lib/Local/GitHub/Make.defs.linux-intel $CHOMBO_HOME/mk/Make.defs.local          
-
-      - name: Build Chombo
-        run: |
-          source /opt/intel/oneapi/setvars.sh          
-          make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} chombo
-
-      - name: Build chombo-discharge/source
-        run: |
-          source /opt/intel/oneapi/setvars.sh
-          make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} discharge-source
-
-      - name: Build chombo-discharge/geometries
-        run: |
-          source /opt/intel/oneapi/setvars.sh
-          make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} discharge-geometries
-
-      - name: Build chombo-discharge/physics
-        run: |
-          source /opt/intel/oneapi/setvars.sh
-          make -s -j${{ env.CORES }} DIM=${{ matrix.dim }} discharge-physics
-
-      - name: Compile tests
-        working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
-        if: always()
-        run: |
-          source /opt/intel/oneapi/setvars.sh          
-          python3 tests.py --compile --clean --no_exec -cores ${{ env.CORES }} -dim ${{ matrix.dim }}
-          if [[ $? == 0 ]]; then exit 0; else exit 1; fi                    
-
-      - name: Run tests
-        working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
-        if: always()        
-        run: |
[…]
