+++ .github/workflows/CI.yml
@@ -157,6 +157,7 @@ jobs:
         working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
         if: always()        
         run: |
+          source /opt/intel/oneapi/setvars.sh                              
           python3 tests.py --compile --clean --no_exec -cores=${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }} -hdf=${{ matrix.hdf }}
           if [[ $? == 0 ]]; then exit 0; else exit 1; fi
 
@@ -164,6 +165,7 @@ jobs:
         working-directory: ${{ env.DISCHARGE_HOME}}/Exec/Tests
         if: always()        
         run: |
+          source /opt/intel/oneapi/setvars.sh                              
           python3 tests.py -cores=${{ env.CORES }} -dim=${{ matrix.dim }} -mpi=${{ matrix.mpi }}
           if [[ $? == 0 ]]; then exit 0; else exit 1; fi                                                  
 
