+++ Docs/Sphinx/source/Base/Installation.rst
@@ -27,28 +27,6 @@ Optional packages are
 * An MPI installation, which is used for parallelization.
 * `<https://visit-dav.github.io/visit-website/>`_ visualization, which used for visualization.
 
-Organization
-------------
-
-The ``chombo-discharge`` source files are organized as follows:
-
-.. list-table:: Code organization.
-   :widths: 10 50
-   :header-rows: 1
-
-   * - Folder
-     - Explanation
-   * - :file:`Source`
-     -  Source files for the AMR core, solvers, and various utilities.
-   * - :file:`Physics`
-     - Various implementations that can run the ``chombo-discharge`` source code.
-   * - :file:`Geometries`
-     - Various geometries.
-   * - :file:`Submodules`
-     - Git submodule dependencies.
-   * - :file:`Exec`
-     - Various executable applications. 
-
 
 Cloning ``chombo-discharge``
 ----------------------------
@@ -71,7 +49,7 @@ In our makefiles the paths to these are supplied through the environment variabl
 Note that ``DISCHARGE_HOME`` must point to the root folder in the ``chombo-discharge`` source code, while ``CHOMBO_HOME`` must point to the :file:`lib/` folder in your ``Chombo`` root directory.
 When cloning with submodules, both ``Chombo`` and ``nlohmann/json`` will be placed in the :file:`Submodules` folder in ``$DISCHARGE_HOME``.  
 
-.. note::
+.. tip::
    
    To clone ``chombo-discharge`` directly to ``$DISCHARGE_HOME``, set the environment variables and clone (using ``--recursive`` to fetch submodules):
 
@@ -121,17 +99,21 @@ If successful, ``chombo-discharge`` libraries will appear in ``$DISCARGE_HOME/Li
 
 .. _Chap:AdvancedConfig:
 
-Advanced configuration
-----------------------
+Full configuration
+------------------
 
-``chombo-discharge`` is compiled using GNU Make, following the ``Chombo``.
-Compilers, libraries, and configuration options are defined in a file ``Make.defs.local`` which resides in ``$CHOMBO_HOME/mk``.
-Users need to supply this file in order to compile ``chombo-discharge``.
-Typically, these steps include
+``chombo-discharge`` is compiled using GNU Make, following the ``Chombo`` configuration methods.
 
-* Specifying Fortran and C++ compilers
+.. important::
+
+   Compilers, libraries, and configuration options are defined in a file ``Make.defs.local`` which resides in ``$CHOMBO_HOME/mk``.
+   Users need to supply this file in order to compile ``chombo-discharge``.
+   
+Typically, a full configuration consists of specifying
+
+* Fortran and C++ compilers
 * Specifying configurations. E.g., serial or parallel builds, and compiler flags. 
-* Specifying library paths (in particular for HDF5).
+* Library paths (in particular for HDF5).
 
 Main settings
 _____________
@@ -150,8 +132,8 @@ The main variables that the user needs to set are
 * ``FC = <Fortran compiler>``
 * ``MPI = TRUE/FALSE``
   This enables/disables MPI.
-* ``MPICXX = <MPI compiler>``
-* ``CXXSTD=14``
+* ``MPICXX = <MPI compiler>`` This sets the MPI compiler.
+* ``CXXSTD = 14`` For specifying the C++ standard. We are currently at C++14.
   Sets the C++ standard - we are currently at C++14.
 * ``USE_EB=TRUE``
   Configures ``Chombo`` with embedded boundary functionality.
@@ -161,6 +143,12 @@ The main variables that the user needs to set are
   This is a requirement.
 * ``USE_HDF5 = TRUE/FALSE``
   This enables and disables HDF5 code.
+
+MPI
+___
+
+To enable MPI, make sure that ``MPI`` is set to true and that the ``MPICXX`` compiler is set.
[…]
