+++ Docs/Sphinx/source/index.rst
@@ -5,7 +5,7 @@
 
 .. important::
 
-   **This is an beta release. Development is still in progress, and various bugs may be present. Minor changes to the user interface can still occur.**
+   **This is a beta release. Development is still in progress, and various bugs may be present. Minor changes to the user interface can still occur.**
 
 Welcome to ``chombo-discharge``'s user documentation
 ****************************************************
@@ -79,7 +79,6 @@ Introduction
    Base/Documentation
    Base/Overview
    Base/Installation
-   Base/Visualization
    Base/Control
    Base/Examples
    Base/Testing
