+++ doxygen/html/search/functions_7.js
@@ -1,11 +1,11 @@
 var searchData=
 [
-  ['harmonicaverage_4713',['harmonicAverage',['../classAmrMesh.html#ae4e9d87d1e36b48c25847153b44b2db2',1,'AmrMesh::harmonicAverage(MFAMRFluxData &amp;a_data, const std::string a_realm) const'],['../classAmrMesh.html#a4b49e5fa25c6a4bb059a6c3e25b1c078',1,'AmrMesh::harmonicAverage(EBAMRCellData &amp;a_data, const std::string a_realm, const phase::which_phase a_phase) const'],['../classAmrMesh.html#adb3e29a7ec2d7e88c8ba7410e1fda2ea',1,'AmrMesh::harmonicAverage(EBAMRCellData &amp;a_data, const std::string a_realm, const phase::which_phase a_phase, const Interval &amp;a_variables) const'],['../classAmrMesh.html#a00b2fdcbb8c36d6ccd8e6b836980ab3c',1,'AmrMesh::harmonicAverage(EBAMRFluxData &amp;a_data, const std::string a_realm, const phase::which_phase a_phase) const'],['../classAmrMesh.html#a259cb2f2336f091eeb1955b7c525a040',1,'AmrMesh::harmonicAverage(EBAMRIVData &amp;a_data, const std::string a_realm, const phase::which_phase a_phase) const'],['../classEBCoarAve.html#a6d27f83ad2c2daf81279a7a22f7041a7',1,'EBCoarAve::harmonicAverage(EBCellFAB &amp;a_coarData, const EBCellFAB &amp;a_fineData, const DataIndex &amp;a_datInd, const Interval &amp;a_coarInterv, const Interval &amp;a_fineInterv)'],['../classEBCoarAve.html#aaeafe5893c90cdd7442e52a640175d5a',1,'EBCoarAve::harmonicAverage(EBFaceFAB &amp;a_coarData, const EBFaceFAB &amp;a_fineData, const DataIndex &amp;a_datInd, const Interval &amp;a_coarInterv, const Interval &amp;a_fineInterv, const int &amp;a_dir)'],['../classEBCoarAve.html#aba84ab32d3c46ecf8d1ba0ee5f072420',1,'EBCoarAve::harmonicAverage(BaseIVFAB&lt; Real &gt; &amp;a_coarData, const BaseIVFAB&lt; Real &gt; &amp;a_fineData, const DataIndex &amp;a_datInd, const Interval &amp;a_coarInterv, const Interval &amp;a_fineInterv) const'],['../classAmrMesh.html#a186c852306b0260ad0637599217ea035',1,'AmrMesh::harmonicAverage()']]],
-  ['hollowcylinderif_4714',['HollowCylinderIF',['../classHollowCylinderIF.html#ab36cf24c8ecd3e78917008c7cf7b83b6',1,'HollowCylinderIF::HollowCylinderIF(const HollowCylinderIF &amp;a_inputIF)'],['../classHollowCylinderIF.html#a29c219c64c244aef4fd0be58d9b30e47',1,'HollowCylinderIF::HollowCylinderIF(const RealVect a_center1, const RealVect a_center2, const Real a_majorRadius, const Real a_minorRadius, const Real a_outerCurvature, const Real a_innerCurvature, const bool a_fluidInside)'],['../classHollowCylinderIF.html#a57cb1c4f8e3f964ea2833bfd28d8f5a0',1,'HollowCylinderIF::HollowCylinderIF()=delete']]],
[…]
