+++ doxygen/html/search/enums_12.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['whichalgorithm_6723',['WhichAlgorithm',['../classPhysics_1_1ItoPlasma_1_1ItoPlasmaGodunovStepper.html#a9ba214fb9cc9d5786d482279e8bd80f7',1,'Physics::ItoPlasma::ItoPlasmaGodunovStepper']]],
-  ['whichcheckpoint_6724',['WhichCheckpoint',['../classItoSolver.html#a93be4b5f935a4ba80472db227c9d75cd',1,'ItoSolver']]],
-  ['whichcontainer_6725',['WhichContainer',['../classItoSolver.html#a070d75d0d3ac16e04013d25574c1ffa5',1,'ItoSolver::WhichContainer()'],['../classMcPhoto.html#a3e74abc2e95a4dd5ff480a4cad3288f6',1,'McPhoto::WhichContainer()']]],
-  ['whichmobilityinterpolation_6726',['WhichMobilityInterpolation',['../classItoSolver.html#ac7ea13986249a06a5185c8315e4548bc',1,'ItoSolver']]]
+  ['velocityfield_6911',['VelocityField',['../namespacePhysics_1_1TracerParticle.html#afc18cec3b9d8a0edb81f7948276a4bc1',1,'Physics::TracerParticle']]]
 ];
