+++ Source/Elliptic/CD_EBHelmholtzEBBC.cpp
@@ -90,7 +90,8 @@ EBHelmholtzEBBC::applyEBFluxResid(VoFIterator&           a_vofit,
                                   const EBCellFAB* const a_phiFine,
                                   const BaseIVFAB<Real>& a_Bcoef,
                                   const DataIndex&       a_dit,
-                                  const Real&            a_beta) const
+                                  const Real&            a_beta,
+                                  const bool&            a_homogeneousPhysBC) const
 {
   CH_TIME("EBHelmholtzEBBC::applyEBFluxResid");
 }
