+++ Source/Elliptic/CD_EBHelmholtzDirichletEBBC.cpp
@@ -203,25 +203,41 @@ EBHelmholtzDirichletEBBC::defineResidualStencils() noexcept
 
   for (DataIterator dit(dbl); dit.ok(); ++dit) {
 
+    BaseIVFAB<Real>&             gradPhiResidWeight       = m_boundaryWeightsResid[dit()];
+    BaseIVFAB<VoFStencil>&       gradPhiResidStencils     = m_gradPhiResidStencils[dit()];
+    BaseIVFAB<VoFStencil>&       gradPhiResidStencilsFine = m_gradPhiResidStencilsFine[dit()];
+    const BaseIVFAB<Real>&       gradPhiRelaxWeight       = m_boundaryWeightsRelax[dit()];
+    const BaseIVFAB<VoFStencil>& gradPhiRelaxStencil      = m_gradPhiRelaxStencils[dit()];
+
+    // Where to define our residual stencils.
+    const IntVectSet& ivs     = gradPhiRelaxStencil.getIVS();
+    const EBISBox&    ebisbox = ebisl[dit()];
+    const EBGraph&    ebgraph = ebisbox.getEBGraph();
+    VoFIterator       vofit(ivs, ebgraph);
+
+    gradPhiResidWeight.define(ivs, ebgraph, m_nComp);
+    gradPhiResidStencils.define(ivs, ebgraph, m_nComp);
+    gradPhiResidStencilsFine.define(ivs, ebgraph, m_nComp);
+
+    // Set default values.
+    auto defaultValues = [&](const VolIndex& vof) -> void {
+      gradPhiResidWeight(vof, m_comp)   = gradPhiRelaxWeight(vof, m_comp);
+      gradPhiResidStencils(vof, m_comp) = gradPhiRelaxStencil(vof, m_comp);
+      gradPhiResidStencilsFine(vof, m_comp).clear();
+    };
+
+    BoxLoops::loop(vofit, defaultValues);
+
     if (m_hasFineAMRLevel && !m_isMGLevel) {
-      BaseIVFAB<Real>&             gradPhiResidWeight       = m_boundaryWeightsResid[dit()];
-      BaseIVFAB<VoFStencil>&       gradPhiResidStencils     = m_gradPhiResidStencils[dit()];
-      BaseIVFAB<VoFStencil>&       gradPhiResidStencilsFine = m_gradPhiResidStencilsFine[dit()];
-      const BaseIVFAB<Real>&       gradPhiRelaxWeight       = m_boundaryWeightsRelax[dit()];
-      const BaseIVFAB<VoFStencil>& gradPhiRelaxStencil      = m_gradPhiRelaxStencils[dit()];
-      const BaseFab<bool>&         validCells               = (*m_amrValidCells)[dit()];
-
-      // Where to define our residual stencils.
-      const IntVectSet& ivs     = gradPhiRelaxStencil.getIVS();
-      const EBGraph&    ebgraph = ebisl[dit()].getEBGraph();
-
-      gradPhiResidWeight.define(ivs, ebgraph, m_nComp);
-      gradPhiResidStencils.define(ivs, ebgraph, m_nComp);
-      gradPhiResidStencilsFine.define(ivs, ebgraph, m_nComp);
+      const BaseFab<bool>& validCells = (*m_amrValidCells)[dit()];
 
       // Go through the relaxation stencils. If it reaches into an invalid cell (which can happen near EBCF),
       // recompute a better stencil for the AMR residual.
       auto kernel = [&](const VolIndex& vof) -> void {
+        gradPhiResidWeight(vof, m_comp)   = gradPhiRelaxWeight(vof, m_comp);
[…]
