+++ Source/Elliptic/CD_EBHelmholtzOp.cpp
@@ -23,8 +23,6 @@
 #include <CD_ParallelOps.H>
 #include <CD_NamespaceHeader.H>
 
-#define USE_OLD_FLUX 0
-
 constexpr int EBHelmholtzOp::m_nComp;
 constexpr int EBHelmholtzOp::m_comp;
 
@@ -678,12 +676,117 @@ EBHelmholtzOp::AMROperator(LevelData<EBCellFAB>&             a_Lphi,
     // If we have a finer level, there is a chance that the stencils from the EB will reach under it, in which case
     // it might fetch potentially bogus data. A clunky way of handling this is to coarsen the data on the fine level
     // first. The best solution would probably be to have the EB flux stencil reach directly into the fine level.
+#if 0
+    // TLDR: This is the original code which does not use the two-sided stencil. 
     if (m_hasFine && m_doCoarsen) {
       EBHelmholtzOp* fineOp = (EBHelmholtzOp*)a_finerOp;
       fineOp->coarsen((LevelData<EBCellFAB>&)a_phi, a_phiFine);
     }
 
     this->applyOp(a_Lphi, a_phi, &a_phiCoar, a_homogeneousPhysBC, false);
+#else
+    // TLDR: This is the new code which uses flux stencils reaching into the fine level also.
+    LevelData<EBCellFAB>& phi = (LevelData<EBCellFAB>&)a_phi;
+
+    if (m_hasFine && m_doCoarsen) {
+      EBHelmholtzOp* fineOp = (EBHelmholtzOp*)a_finerOp;
+      fineOp->coarsen((LevelData<EBCellFAB>&)a_phi, a_phiFine);
+    }
+
+    if (m_doExchange) {
+      phi.exchange();
+    }
+
+    if (m_hasCoar && m_doInterpCF) {
+      this->interpolateCF(phi, &a_phiCoar, false);
+    }
+
+    if (m_hasFine) {
+      a_phiFine.copyTo(m_phiFine);
+    }
+
+    // Apply operator in each kernel.
+    const DisjointBoxLayout& dbl = a_Lphi.disjointBoxLayout();
+    for (DataIterator dit(dbl); dit.ok(); ++dit) {
+      const Box&     cellBox = m_eblg.getDBL()[dit()];
+      const EBISBox& ebisbox = m_eblg.getEBISL()[dit()];
+
+      EBCellFAB& Lphi = a_Lphi[dit()];
+
+      if (!ebisbox.isAllCovered()) {
+        this->applyOpRegular(Lphi, phi[dit()], cellBox, dit(), a_homogeneousPhysBC);
+
+        VoFIterator& vofit = m_vofIterStenc[dit()];
+        for (vofit.reset(); vofit.ok(); ++vofit) {
+          const VolIndex& vof = vofit();
+
+          // Finite volume stencil representation of Int[B*grad(phi)]dA
+          const VoFStencil& stenc = m_relaxStencils[dit()](vof, m_comp);
+
+          // kappa * alpha * aco (m_alphaDiagWeight holds kappa* aco)
+          const Real& alphaDiag = m_alpha * m_alphaDiagWeight[dit()](vof, m_comp);
+
+          a_Lphi[dit()](vof, m_comp) = alphaDiag * a_phi[dit()](vof, m_comp);
+
+          // Add the term corresponding to kappa*div(b*grad(phi)), excluding contributions from domain faces.
+          for (int i = 0; i < stenc.size(); i++) {
+            const VolIndex& ivof    = stenc.vof(i);
+            const Real&     iweight = stenc.weight(i);
+
+            // Note that bco is a part of the stencil weight.
+            a_Lphi[dit()](vof, m_comp) += m_beta * iweight * a_phi[dit()](ivof, m_comp);
+          }
+        }
+
[…]
