### 2023-04-12 25874ba
Starting to figure out the discretization and what the signatures need to look like. Park for now.
