+++ Source/Elliptic/CD_EBHelmholtzDirichletEBBC.cpp
@@ -393,6 +393,16 @@ EBHelmholtzDirichletEBBC::applyEBFluxResid(VoFIterator&           a_vofit,
       DphiDn += iweight * a_phi(ivof, m_comp);
     }
 
+#if 1 // Temporary hack because of how EBHelmholtzOp stores the cut-cell divOp stencil
+    const VoFStencil& st2 = m_gradPhiRelaxStencils[a_dit](vof, m_comp);
+    for (int i = 0; i < st2.size(); i++) {
+      const VolIndex& ivof    = st2.vof(i);
+      const Real&     iweight = st2.weight(i);
+
+      DphiDn -= iweight * a_phi(ivof, m_comp);
+    }
+#endif
+
     if (m_hasFineAMRLevel && !m_isMGLevel) {
       const EBCellFAB& phiFine = *a_phiFine;
 
