+++ Source/Elliptic/CD_EBHelmholtzOp.cpp
@@ -263,7 +263,7 @@ EBHelmholtzOp::getFlux() const
 void
 EBHelmholtzOp::defineStencils()
 {
-  CH_TIME("EBHelmholtzOp::getFlux()");
+  CH_TIME("EBHelmholtzOp::defineStencils()");
 
   // Basic defines.
   EBCellFactory cellFact(m_eblg.getEBISL());
@@ -278,6 +278,7 @@ EBHelmholtzOp::defineStencils()
   m_alphaDiagWeight.define(m_eblg.getDBL());
   m_betaDiagWeight.define(m_eblg.getDBL());
   m_relaxStencils.define(m_eblg.getDBL());
+  m_opStencilsNoEB.define(m_eblg.getDBL());
 
   for (int dir = 0; dir < SpaceDim; dir++) {
     m_vofIterDomLo[dir].define(m_eblg.getDBL());
@@ -360,6 +361,7 @@ EBHelmholtzOp::defineStencils()
 
     // Define data holders.
     m_relaxStencils[dit()].define(stencIVS, ebgraph, m_nComp);
+    m_opStencilsNoEB[dit()].define(stencIVS, ebgraph, m_nComp);
     m_alphaDiagWeight[dit()].define(stencIVS, ebgraph, m_nComp);
     m_betaDiagWeight[dit()].define(stencIVS, ebgraph, m_nComp);
 
@@ -401,12 +403,14 @@ EBHelmholtzOp::defineStencils()
     // Referring to the sketch above, we iterate through cells A, B, D, E and compute the face centroid fluxes for all these cells. This includes
     // e.g. the face connecting E and F. However, since F only has regular faces we don't store the stencil explicitly for that cell.
     //
-    BaseIVFAB<VoFStencil>& opStencil  = m_relaxStencils[dit()];
-    VoFIterator&           vofitStenc = m_vofIterStenc[dit()];
-    VoFIterator&           vofitIrreg = m_vofIterIrreg[dit()];
+    BaseIVFAB<VoFStencil>& relaxStencil  = m_relaxStencils[dit()];
+    BaseIVFAB<VoFStencil>& opStencilNoEB = m_opStencilsNoEB[dit()];
+    VoFIterator&           vofitStenc    = m_vofIterStenc[dit()];
+    VoFIterator&           vofitIrreg    = m_vofIterIrreg[dit()];
 
     BoxLoops::loop(vofitStenc, [&](const VolIndex& vof) -> void {
-      opStencil(vof, m_comp).clear();
+      relaxStencil(vof, m_comp).clear();
+      opStencilNoEB(vof, m_comp).clear();
     });
 
     for (int dir = 0; dir < SpaceDim; dir++) {
@@ -438,10 +442,12 @@ EBHelmholtzOp::defineStencils()
 
           // 4. Note that we prune storage here.
           if (stencIVS.contains(vofLo.gridIndex())) {
-            opStencil(vofLo, m_comp) += loKappaDivFSten;
+            relaxStencil(vofLo, m_comp) += loKappaDivFSten;
+            opStencilNoEB(vofLo, m_comp) += loKappaDivFSten;
           }
           if (stencIVS.contains(vofHi.gridIndex())) {
-            opStencil(vofHi, m_comp) += hiKappaDivFSten;
+            relaxStencil(vofHi, m_comp) += hiKappaDivFSten;
+            opStencilNoEB(vofHi, m_comp) += hiKappaDivFSten;
           }
         }
       };
@@ -455,7 +461,7 @@ EBHelmholtzOp::defineStencils()
 
       ebSten *= (*m_BcoefIrreg)[dit()](vof, m_comp);
 
-      opStencil(vof, m_comp) += ebSten;
+      relaxStencil(vof, m_comp) += ebSten;
     });
   }
 
[…]
