+++ Physics/StreamerInception/CD_StreamerInceptionStepperImplem.H
@@ -1057,22 +1057,22 @@ StreamerInceptionStepper<P, F, C>::computeDt()
       break;
     }
     }
-    auto computeGradient = [this](const Real & a_timeStep){
-			     return (m_voltageCurve(m_time+a_timeStep)-m_voltageCurve(m_time))/ a_timeStep;
-			   };
-    Real prevDt = m_minDt;
-    Real currDt = 2*m_minDt;
-    Real currGradient = computeGradient(currDt);
-    const Real currVoltage = m_voltageCurve(m_time);
-    Real deltaV = currGradient*currDt;
-    while(std::abs(deltaV) <= std::abs(m_deltaVPercent*currVoltage) && (m_time+currDt) < m_endTime){
+    auto computeGradient = [this](const Real& a_timeStep) {
+      return (m_voltageCurve(m_time + a_timeStep) - m_voltageCurve(m_time)) / a_timeStep;
+    };
+    Real       prevDt       = m_minDt;
+    Real       currDt       = 2 * m_minDt;
+    Real       currGradient = computeGradient(currDt);
+    const Real currVoltage  = m_voltageCurve(m_time);
+    Real       deltaV       = currGradient * currDt;
+    while (std::abs(deltaV) <= std::abs(m_deltaVPercent * currVoltage) && (m_time + currDt) < m_endTime) {
       prevDt = currDt;
       currDt += m_minDt;
       currGradient = computeGradient(currDt);
-      deltaV = currGradient*currDt;
+      deltaV       = currGradient * currDt;
     }
     dt = prevDt;
-    
+
     dt = std::min(dt, m_maxDt);
     dt = std::max(dt, m_minDt);
   }
