+++ Source/Particle/CD_ParticleManagementImplem.H
@@ -472,9 +472,7 @@ namespace ParticleManagement {
 
   template <typename P, typename T, typename>
   inline void
-  drawRandomParticles(List<P>&                         a_particles,
-                                          const T                          a_numParticles,
-                                          const std::function<RealVect()>& a_distribution)
+  drawRandomParticles(List<P>& a_particles, const T a_numParticles, const std::function<RealVect()>& a_distribution)
   {
 
     a_particles.clear();
@@ -492,9 +490,9 @@ namespace ParticleManagement {
   template <typename P, typename T, typename>
   inline void
   drawSphereParticles(List<P>&       a_particles,
-                                          const T        a_numParticles,
-                                          const RealVect a_center,
-                                          const Real     a_radius) noexcept
+                      const T        a_numParticles,
+                      const RealVect a_center,
+                      const Real     a_radius) noexcept
   {
     CH_TIME("ParticleManagement::drawSphereParticles");
 
@@ -524,9 +522,9 @@ namespace ParticleManagement {
   template <typename P, typename T, typename>
   inline void
   drawBoxParticles(List<P>&       a_particles,
-                                       const T        a_numParticles,
-                                       const RealVect a_loCorner,
-                                       const RealVect a_hiCorner) noexcept
+                   const T        a_numParticles,
+                   const RealVect a_loCorner,
+                   const RealVect a_hiCorner) noexcept
   {
     CH_TIME("ParticleManagement::drawBoxParticles");
 
@@ -544,9 +542,9 @@ namespace ParticleManagement {
   template <typename P, typename T, typename>
   inline void
   drawGaussianParticles(List<P>&       a_particles,
-                                            const T        a_numParticles,
-                                            const RealVect a_center,
-                                            const Real     a_radius) noexcept
+                        const T        a_numParticles,
+                        const RealVect a_center,
+                        const Real     a_radius) noexcept
   {
     CH_TIME("ParticleManagement::drawGaussianParticles");
 
