### 2023-04-14 ccdf1d3
added computeIonizationVolume-function for both stationary and transient mode. Not finished
