+++ Physics/StreamerInception/CD_StreamerInceptionStepperImplem.H
@@ -2896,6 +2896,169 @@ StreamerInceptionStepper<P, F, C>::computeCriticalVolumeTransient() const noexce
 
 template <typename P, typename F, typename C>
 void
+StreamerInceptionStepper<P, F, C>::computeIonizationVolumeStationary() noexcept
+{
+  CH_TIME("StreamerInceptionStepper::computeIonizationVolumeStationary");
+  if (m_verbosity > 5) {
+    pout() << "StreamerInceptionStepper::computeIonizationVolumeStationary" << endl;
+  }
+
+  // Solve ionization volume for each voltage
+  for (size_t i = 0; i < m_voltageSweeps.size(); ++i) {
+    Real ionizationVolumePlus = 0.0;
+    Real ionizationVolumeMinu = 0.0;
+
+    for (int lvl = 0; lvl <= m_amr->getFinestLevel(); ++lvl) {
+      const DisjointBoxLayout& dbl   = m_amr->getGrids(m_realm)[lvl];
+      const EBISLayout&        ebisl = m_amr->getEBISLayout(m_realm, m_phase)[lvl];
+
+      const LevelData<BaseFab<bool>>& validCellsLD = *m_amr->getValidCells(m_realm)[lvl];
+
+      const Real dx = m_amr->getDx()[lvl];
+
+      for (DataIterator dit = dbl.dataIterator(); dit.ok(); ++dit) {
+        const EBISBox&       ebisbox    = ebisl[dit()];
+        const BaseFab<bool>& validCells = validCellsLD[dit()];
+
+        const EBCellFAB& electricFieldPlus = (*m_electricFieldPlus[lvl])[dit()];
+        const EBCellFAB& electricFieldMinu = (*m_electricFieldMinu[lvl])[dit()];
+
+        const FArrayBox& electricFieldRegPlus = electricFieldPlus.getFArrayBox();
+        const FArrayBox& electricFieldRegMinu = electricFieldMinu.getFArrayBox();
+
+        auto regularKernel = [&](const IntVect& iv) -> void {
+          if (validCells(iv, 0) && ebisbox.isRegular(iv)) {
+            const RealVect EEplus = RealVect(
+              D_ECL(electricFieldRegPlus(iv, 0), electricFieldRegPlus(iv, 1), electricFieldRegPlus(iv, 2)));
+            const RealVect EEminu = RealVect(
+              D_ECL(electricFieldRegMinu(iv, 0), electricFieldRegMinu(iv, 1), electricFieldRegMinu(iv, 2)));
+
+            const Real Eplus = std::abs(1 * EEplus.vectorLength()); // change 1 with the voltage
+            const Real Eminu = std::abs(1 * EEminu.vectorLength()); // change 1 with the voltage
+
+            const Real alphaPlus = m_alpha(Eplus);
+            const Real alphaMinu = m_alpha(Eminu);
+
+            const Real etaPlus = m_eta(Eplus);
+            const Real etaMinu = m_eta(Eminu);
+            if (alphaPlus >= etaPlus) {
+              ionizationVolumePlus += std::pow(dx, SpaceDim);
+            }
+            if (alphaMinu >= etaMinu) {
+              ionizationVolumeMinu += std::pow(dx, SpaceDim);
+            }
+          }
+        };
+
+        auto irregularKernel = [&](const VolIndex& vof) -> void {
+          if (validCells(vof.gridIndex())) {
+            const RealVect EEplus = RealVect(
+              D_ECL(electricFieldPlus(vof, 0), electricFieldPlus(vof, 1), electricFieldPlus(vof, 2)));
+            const RealVect EEminu = RealVect(
[…]
