### 2023-04-17 4e2e29f
fixed a few bugs + clang-format
