+++ Physics/StreamerInception/CD_StreamerInceptionStepper.H
@@ -482,15 +482,9 @@ namespace Physics {
 
       /*!
 	@brief Ionization volumes for each voltage
-	@note Stationary mode, positive polarity
+	@note Stationary mode
       */
-      std::vector<Real> m_ionizationVolumePlus;
-
-      /*!
-	@brief Ionization volumes for each voltage
-	@note Stationary mode, negative polarity
-      */
-      std::vector<Real> m_ionizationVolumeMinu;
+      std::vector<Real> m_ionizationVolume;
 
       /*!
 	@brief Rate of appearance of first electron (in critical volume)
