### 2023-04-19 fafadbe
added computeIonizationVolumeStationary() call in postInitialize()
