+++ Physics/TracerParticle/CD_TracerParticlePhysics.H
@@ -18,7 +18,8 @@ namespace Physics {
   /*!
     @brief Namespace for encapsulating physics code for tracer particles
   */
-  namespace TracerParticle {}
+  namespace TracerParticle {
+  }
 
 } // namespace Physics
 
