+++ Physics/BrownianWalker/CD_BrownianWalker.H
@@ -17,7 +17,8 @@ namespace Physics {
   /*!
     @brief Namespace for encapsulating physics code for describing Brownian walkers. 
   */
-  namespace BrownianWalker {}
+  namespace BrownianWalker {
+  }
 
 } // namespace Physics
 
