+++ Physics/AdvectionDiffusion/CD_AdvectionDiffusion.H
@@ -16,7 +16,8 @@ namespace Physics {
   /*!
     @brief Namespace for encapsulating AdvectionDiffusion physics code.
   */
-  namespace AdvectionDiffusion {};
+  namespace AdvectionDiffusion {
+  };
 
 } // namespace Physics
 
