+++ Physics/CD_Physics.H
@@ -15,6 +15,7 @@
 /*!
   @brief Name containing various physics models for running chombo-discharge code
 */
-namespace Physics {};
+namespace Physics {
+};
 
 #endif
