+++ Physics/CdrPlasma/CD_CdrPlasma.H
@@ -16,7 +16,8 @@ namespace Physics {
   /*!
     @brief Namespace for encapsulating CDR plasma code.
   */
-  namespace CdrPlasma {};
+  namespace CdrPlasma {
+  };
 
 } // namespace Physics
 
