### 2023-04-21 b38d6ab
included computeCriticalArea for both stationary and transient
