+++ Physics/StreamerInception/CD_StreamerInceptionStepperImplem.H
@@ -1108,12 +1108,14 @@ StreamerInceptionStepper<P, F, C>::advance(const Real a_dt)
     this->seedIonizationParticles(curVoltage);
     this->computeInceptionIntegralTransient(curVoltage);
 
-    // Compute the critical volume, the ionization volume and the electron appearance rate.
+    // Compute the critical volume, critical area, the ionization volume and the electron appearance rate.
     const Real Vcr  = this->computeCriticalVolumeTransient();
+    const Real Acr = this->computeCriticalAreaTransient();
     const Real Vion = this->computeIonizationVolumeTransient(curVoltage);
     const Real Rdot = this->computeRdot(curVoltage);
 
     m_criticalVolume.emplace_back(curTime, Vcr);
+    m_criticalArea.emplace_back(curTime, Acr);
     m_ionizationVolumeTransient.emplace_back(curTime, Vion);
     m_Rdot.emplace_back(curTime, Rdot);
 
@@ -1701,6 +1703,7 @@ StreamerInceptionStepper<P, F, C>::postInitialize()
     this->computeInceptionIntegralStationary();
     this->computeInceptionVoltageVolume();
     this->computeCriticalVolumeStationary();
+    this->computeCriticalAreaStationary();
     this->computeIonizationVolumeStationary();
     this->computeRdotStationary();
 
@@ -1738,6 +1741,7 @@ StreamerInceptionStepper<P, F, C>::postInitialize()
 
     m_Rdot.emplace_back(m_time, this->computeRdot(m_voltageCurve(m_time)));
     m_criticalVolume.emplace_back(m_time, this->computeCriticalVolumeTransient());
+    m_criticalArea.emplace_back(m_time, this->computeCriticalAreaTransient());
     m_ionizationVolumeTransient.emplace_back(m_time, this->computeIonizationVolumeTransient(m_voltageCurve(m_time)));
     m_inceptionProbability.emplace_back(m_time, 0.0);
     m_maxK.emplace_back(m_time, maxK);
@@ -2899,6 +2903,126 @@ StreamerInceptionStepper<P, F, C>::computeCriticalVolumeTransient() const noexce
   return ParallelOps::sum(Vcr);
 }
 
+template<typename P, typename F, typename C>
+void
+StreamerInceptionStepper<P, F, C>::computeCriticalAreaStationary() noexcept {
+  CH_TIME("StreamerInceptionStepper::computeCriticalAreaStationary");
+  if(m_verbosity > 5){
+    pout() << "StreamerInceptionStepper::computeCriticalAreaStationary" << endl;
+  }
+
+  // Solve critical area of K values for each voltage
+  for (size_t i = 0; i < m_voltageSweeps.size(); ++i){
+    Real criticalAreaPlus = 0.0;
+    Real criticalAreaMinu = 0.0;
+
+    for(int lvl = 0; lvl <= m_amr->getFinestLevel(); ++lvl){
+      const DisjointBoxLayout& dbl = m_amr->getGrids(m_realm)[lvl];
+      const EBISLayout& ebisl = m_amr->getEBISLayout(m_realm, m_phase)[lvl];
+
+      const LevelData<BaseFab<bool>>& validCellsLD = *m_amr->getValidCells(m_realm)[lvl];
+
+      const Real dx = m_amr->getDx()[lvl];
+
+      for(DataIterator dit = dbl.dataIterator(); dit.ok(); ++dit){
+	const EBISBox& ebisbox = ebisl[dit()];
+	const BaseFab<bool>& validCells = validCellsLD[dit()];
+
[…]
