### 2023-04-21 01fe952
Copy stencils from jump bc class
