### 2023-04-22 4f30daa
Define aggstencils now
