+++ Source/Elliptic/CD_MFHelmholtzJumpBC.cpp
@@ -49,6 +49,9 @@ MFHelmholtzJumpBC::MFHelmholtzJumpBC(const Location::Cell a_dataLocation,
   m_numPhases    = m_mflg.numPhases();
   m_multiPhase   = m_numPhases > 1;
 
+  m_ghostPhi = 2 * IntVect::Unit;
+  MayDay::Warning("MFHelmholtzJumpBC -- must get number of ghost cells for using AggStencil properly");
+
   this->defineIterators();
   this->defineStencils();
 }
@@ -317,11 +320,7 @@ MFHelmholtzJumpBC::buildAverageStencils()
     aggStencils.define(dbl);
 
     // Some dummy data for the stencils.
-    EBCellFactory       cellFact(ebisl);
-    BaseIVFactory<Real> irregFact(ebisl);
-
-    LevelData<EBCellFAB>       phiProxy(dbl, 1, 2 * IntVect::Unit, cellFact);
-    LevelData<BaseIVFAB<Real>> jumpProxy(dbl, 1, IntVect::Zero, irregFact);
+    LevelData<EBCellFAB> phiProxy(dbl, 1, m_ghostPhi, EBCellFactory(ebisl));
 
     for (DataIterator dit(dbl); dit.ok(); ++dit) {
       const EBISBox& ebisbox = ebisl[dit()];
@@ -342,7 +341,7 @@ MFHelmholtzJumpBC::buildAverageStencils()
       BoxLoops::loop(vofit, kernel);
 
       aggStencils[dit()] = RefCountedPtr<AggStencil<EBCellFAB, BaseIVFAB<Real>>>(
-        new AggStencil(dstBaseIndex, dstBaseStencil, phiProxy[dit()], jumpProxy[dit()]));
+        new AggStencil(dstBaseIndex, dstBaseStencil, phiProxy[dit()], m_boundaryPhi[dit()].getIVFAB(iphase)));
     }
   }
 }
@@ -472,11 +471,11 @@ MFHelmholtzJumpBC::matchBC(BaseIVFAB<Real>& a_jump,
   CH_assert(m_multiPhase);
 
   CH_TIMERS("MFHelmholtzJumpBC::matchBC(patch)");
-  CH_TIMER("one", t1);
-  CH_TIMER("two", t2);
-  CH_TIMER("three", t3);
-  CH_TIMER("four", t4);
-  CH_TIMER("five", t5);
+  CH_TIMER("Setup", t1);
+  CH_TIMER("Apply stencils", t2);
+  CH_TIMER("Get vofs", t3);
+  CH_TIMER("Compute jump", t4);
+  CH_TIMER("Compute phiBndry", t5);
 
   // TLDR: This routine computes the boundary value of phi from an expression b1*dphi/dn1 + b2*dphi/dn2 = sigma, where dphi/dn can be represented as
   //
@@ -487,7 +486,7 @@ MFHelmholtzJumpBC::matchBC(BaseIVFAB<Real>& a_jump,
   //           phiB = sigma/(b1*w1 + b2*w2) - b1*sum[w1(i) * phi1(i)]/(b1*w1 + b2*w2) - b2*sum[w2(i) * phi2(i)]/(b1*w1 + b2*w2).
   //
   //       Because I'm not crazy, I have stored the term 1/(b1*w1 + b2*w2) in m_denom so we can just multiply it in when we need it. Moreover,
-  //       this term has already been multiplied into the stencil weights, which is the reason why we only do applyStencil below (without dividing
[…]
