### 2023-04-22 51b66aa
Some notes for myself
