### 2023-04-22 e1ccd23
More optimizations
