+++ Source/Elliptic/CD_MFHelmholtzJumpBC.cpp
@@ -30,7 +30,8 @@ MFHelmholtzJumpBC::MFHelmholtzJumpBC(const Location::Cell a_dataLocation,
                                      const int            a_order,
                                      const int            a_weight,
                                      const int            a_radius,
-                                     const int            a_ghostCF)
+                                     const int            a_ghostCF,
+                                     const IntVect        a_ghostPhi)
 {
   CH_TIME("MFHelmholtzJumpBC::MFHelmholtzJumpBC");
 
@@ -48,9 +49,8 @@ MFHelmholtzJumpBC::MFHelmholtzJumpBC(const Location::Cell a_dataLocation,
   m_ghostCF      = a_ghostCF;
   m_numPhases    = m_mflg.numPhases();
   m_multiPhase   = m_numPhases > 1;
+  m_ghostPhi     = a_ghostPhi;
 
-  m_ghostPhi = 2 * IntVect::Unit;
-  MayDay::Warning("MFHelmholtzJumpBC -- must get number of ghost cells for using AggStencil properly");
   MayDay::Warning("MFHelmholtzJumpBC -- should store target vofs for average jump calc");
   MayDay::Warning("MFHelmholtzJumpBC -- check if we can trim exchange operation");
 
@@ -140,6 +140,7 @@ MFHelmholtzJumpBC::defineStencils()
     m_boundaryPhi.define(dbl);
     m_avgStencils.define(dbl);
     m_avgWeights.define(dbl);
+    m_avgVoFs.define(dbl);
     m_denom.define(dbl);
 
     for (DataIterator dit(dbl); dit.ok(); ++dit) {
@@ -151,6 +152,7 @@ MFHelmholtzJumpBC::defineStencils()
       m_avgStencils[dit()].define(m_mflg, dit());
       m_avgWeights[dit()].define(m_mflg, dit());
       m_denom[dit()].define(m_mflg, dit());
+      m_avgVoFs[dit()].define(m_mflg, dit());
     }
 
     // Build stencils and weights for each phase.
@@ -165,8 +167,8 @@ MFHelmholtzJumpBC::defineStencils()
         const IntVectSet& ivs     = m_ivs[dit()];
 
         // Build stencils like we always do
-        BaseIVFAB<VoFStencil>& gradStencils = m_gradPhiStencils[dit()].getIVFAB(iphase); //(ivs, ebgraph, m_nComp);
-        BaseIVFAB<Real>&       bndryWeights = m_gradPhiWeights[dit()].getIVFAB(iphase);  //(ivs, ebgraph, m_nComp);
+        BaseIVFAB<VoFStencil>& gradStencils = m_gradPhiStencils[dit()].getIVFAB(iphase);
+        BaseIVFAB<Real>&       bndryWeights = m_gradPhiWeights[dit()].getIVFAB(iphase);
 
         // Iteration space for kernel
         VoFIterator vofit(ivs, ebgraph);
@@ -247,9 +249,10 @@ MFHelmholtzJumpBC::buildAverageStencils()
       const BaseIVFAB<Real>&       bndryWeights = m_gradPhiWeights[dit()].getIVFAB(iphase);
 
       // Build the average stencils. Only matters if the cell is a multi-valued cell.
[…]
