+++ Source/Electrostatics/CD_MFHelmholtzSaturationChargeJumpBC.H
@@ -37,6 +37,7 @@ public:
     @param[in] a_weight       Weighting factor for least squares
     @param[in] a_radius       Stencil radius
     @param[in] a_ghostCF      Number of grid cells that were filled across the CF. 
+    @param[in] a_ghostPhi     Number of ghost cells in phi (needed for AggStencil)
   */
   MFHelmholtzSaturationChargeJumpBC(const phase::which_phase a_phase,
                                     const Location::Cell     a_dataLocation,
@@ -46,12 +47,13 @@ public:
                                     const int                a_order,
                                     const int                a_weight,
                                     const int                a_radius,
-                                    const int                a_ghostCF);
+                                    const int                a_ghostCF,
+				    const IntVect            a_ghostPhi);
 
   /*!
     @brief Destructor (does nothing)
   */
-  ~MFHelmholtzSaturationChargeJumpBC();
+  virtual ~MFHelmholtzSaturationChargeJumpBC();
 
   /*!
     @brief Match the BC. 
