### 2023-04-23 bcccc78
Optimizations for linear solvers (#327)
