+++ Source/Elliptic/CD_MFHelmholtzJumpBC.cpp
@@ -9,6 +9,10 @@
   @author Robert Marskar
 */
 
+// Chombo includes
+#include <EBCellFactory.H>
+#include <BaseIVFactory.H>
+
 // Our includes
 #include <CD_MFHelmholtzJumpBC.H>
 #include <CD_VofUtils.H>
@@ -26,7 +30,8 @@ MFHelmholtzJumpBC::MFHelmholtzJumpBC(const Location::Cell a_dataLocation,
                                      const int            a_order,
                                      const int            a_weight,
                                      const int            a_radius,
-                                     const int            a_ghostCF)
+                                     const int            a_ghostCF,
+                                     const IntVect        a_ghostPhi)
 {
   CH_TIME("MFHelmholtzJumpBC::MFHelmholtzJumpBC");
 
@@ -44,6 +49,7 @@ MFHelmholtzJumpBC::MFHelmholtzJumpBC(const Location::Cell a_dataLocation,
   m_ghostCF      = a_ghostCF;
   m_numPhases    = m_mflg.numPhases();
   m_multiPhase   = m_numPhases > 1;
+  m_ghostPhi     = a_ghostPhi;
 
   this->defineIterators();
   this->defineStencils();
@@ -87,6 +93,18 @@ MFHelmholtzJumpBC::getMultiPhaseVofs(const int a_phase, const DataIndex& a_dit) 
   return (*m_multiPhaseVofs.at(a_phase))[a_dit];
 }
 
+const LayoutData<MFInterfaceFAB<VoFStencil>>&
+MFHelmholtzJumpBC::getGradPhiStencils() const noexcept
+{
+  return m_gradPhiStencils;
+}
+
+const LayoutData<MFInterfaceFAB<Real>>&
+MFHelmholtzJumpBC::getGradPhiWeights() const noexcept
+{
+  return m_gradPhiWeights;
+}
+
 void
 MFHelmholtzJumpBC::setBco(const RefCountedPtr<LevelData<MFBaseIVFAB>>& a_Bcoef)
 {
@@ -97,6 +115,13 @@ MFHelmholtzJumpBC::setBco(const RefCountedPtr<LevelData<MFBaseIVFAB>>& a_Bcoef)
   if (m_multiPhase) {
     this->buildAverageStencils();
   }
+}
+bool
+MFHelmholtzJumpBC::isMultiPhase() const noexcept
+{
+  CH_TIME("MFHelmholtzJumpBC::isMultiPhase");
+
+  return m_multiPhase;
 }
 
 void
@@ -119,6 +144,7 @@ MFHelmholtzJumpBC::defineStencils()
     m_boundaryPhi.define(dbl);
     m_avgStencils.define(dbl);
     m_avgWeights.define(dbl);
+    m_avgVoFs.define(dbl);
     m_denom.define(dbl);
 
     for (DataIterator dit(dbl); dit.ok(); ++dit) {
@@ -130,6 +156,7 @@ MFHelmholtzJumpBC::defineStencils()
       m_avgStencils[dit()].define(m_mflg, dit());
       m_avgWeights[dit()].define(m_mflg, dit());
       m_denom[dit()].define(m_mflg, dit());
+      m_avgVoFs[dit()].define(m_mflg, dit());
     }
 
     // Build stencils and weights for each phase.
@@ -144,8 +171,8 @@ MFHelmholtzJumpBC::defineStencils()
         const IntVectSet& ivs     = m_ivs[dit()];
 
         // Build stencils like we always do
[…]
