+++ doxygen/html/search/enums_b.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['mode_6904',['Mode',['../CD__StreamerInceptionStepper_8H.html#af655f05a634af7f120b02f162dc8fa8f',1,'Physics::StreamerInception']]],
-  ['multigridtype_6905',['MultigridType',['../classCdrMultigrid.html#a004b5dca8c7ba07c58383012460132b8',1,'CdrMultigrid::MultigridType()'],['../classFieldSolverMultigrid.html#aaf89c3ef8e19504afb28976473f2b797',1,'FieldSolverMultigrid::MultigridType()'],['../classEddingtonSP1.html#af23167e2e38a68e6c3016045fe2c8a9d',1,'EddingtonSP1::MultigridType()']]]
+  ['mode_6909',['Mode',['../CD__StreamerInceptionStepper_8H.html#af655f05a634af7f120b02f162dc8fa8f',1,'Physics::StreamerInception']]],
+  ['multigridtype_6910',['MultigridType',['../classCdrMultigrid.html#a004b5dca8c7ba07c58383012460132b8',1,'CdrMultigrid::MultigridType()'],['../classFieldSolverMultigrid.html#aaf89c3ef8e19504afb28976473f2b797',1,'FieldSolverMultigrid::MultigridType()'],['../classEddingtonSP1.html#af23167e2e38a68e6c3016045fe2c8a9d',1,'EddingtonSP1::MultigridType()']]]
 ];
