+++ doxygen/html/search/typedefs_6.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['initfunction_6861',['InitFunction',['../classPhysics_1_1CdrPlasma_1_1CdrSpeciesJSON.html#a746601a091d29db7a9150bbf1527bb24',1,'Physics::CdrPlasma::CdrSpeciesJSON']]],
-  ['initialdatafunction_6862',['InitialDataFunction',['../classPhysics_1_1CdrPlasma_1_1CdrPlasmaJSON.html#a1b3c8227a830f100dfe51cac50a4c43c',1,'Physics::CdrPlasma::CdrPlasmaJSON']]]
+  ['initfunction_6866',['InitFunction',['../classPhysics_1_1CdrPlasma_1_1CdrSpeciesJSON.html#a746601a091d29db7a9150bbf1527bb24',1,'Physics::CdrPlasma::CdrSpeciesJSON']]],
+  ['initialdatafunction_6867',['InitialDataFunction',['../classPhysics_1_1CdrPlasma_1_1CdrPlasmaJSON.html#a1b3c8227a830f100dfe51cac50a4c43c',1,'Physics::CdrPlasma::CdrPlasmaJSON']]]
 ];
