### 2023-04-24 2f1aa93
fixed wrong calculation in computeCriticalArea
