### 2023-04-25 63d15f6
Started adding buffers and letting operators use them. Want to shrink memory
