### 2023-04-25 32f86a7
WIP on simplifying multigrid interpolator
