+++ Source/AmrMesh/CD_EBLeastSquaresMultigridInterpolator.cpp
@@ -33,21 +33,28 @@ constexpr int EBLeastSquaresMultigridInterpolator::m_numStenComp;
 constexpr int EBLeastSquaresMultigridInterpolator::m_comp;
 
 EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator(const EBLevelGrid& a_eblgFine,
+                                                                         const EBLevelGrid& a_eblgCoFi,
                                                                          const EBLevelGrid& a_eblgCoar,
                                                                          const CellLocation a_dataLocation,
                                                                          const IntVect&     a_ghostVector,
                                                                          const int          a_refRat,
                                                                          const int          a_ghostCF,
                                                                          const int          a_order,
-                                                                         const int          a_weighting)
+                                                                         const int          a_weighting) noexcept
 {
-  CH_TIME("EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator");
+  CH_TIMERS("EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator");
+  CH_TIMER("define_regular", t1);
+  CH_TIMER("define_irregular", t2);
 
   CH_assert(a_ghostCF > 0);
   CH_assert(a_refRat % 2 == 0);
+  CH_assert(a_eblgFine.getGhost() >= a_ghostVector.max());
 
-  const DisjointBoxLayout& gridsFine = a_eblgFine.getDBL();
-  const DisjointBoxLayout& gridsCoar = a_eblgCoar.getDBL();
+  // Not enough ghost cells in input eblg. I don't know why anyone would do that, but make sure we abort if it happens. This is a more
+  // transparent error message than the assertion.
+  if (a_eblgFine.getGhost() < a_ghostVector.max()) {
+    MayDay::Error("EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator - not enough ghost cells!");
+  }
 
   // Build the regular stencil objects for regular-grid interpolation. For now, we use QuadCFInterp for this. I also leave
   // a timer in place until performance and scalability has been investigated. To check performance, add
@@ -60,19 +67,22 @@ EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator(const E
   m_ghostVector  = a_ghostVector;
   m_dataLocation = a_dataLocation;
   m_weight       = a_weighting;
+  m_eblgFine     = a_eblgFine;
+  m_eblgCoFi     = a_eblgCoFi;
+  m_eblgCoar     = a_eblgCoar;
 
+  // Build the regular stencil objects for regular-grid interpolation. Use QuadCFInterp for this (might change this later)
+  CH_START(t1);
   timer.startEvent("Define QuadCFInterp");
-  //  QuadCFInterp::define(gridsFine, &gridsCoar, 1.0, a_refRat, 1, a_eblgFine.getDomain());
-  m_scalarInterpolator = RefCountedPtr<QuadCFInterp>(
+  const DisjointBoxLayout& gridsFine = a_eblgFine.getDBL();
+  const DisjointBoxLayout& gridsCoar = a_eblgCoar.getDBL();
+
+  m_quadCFInterp = RefCountedPtr<QuadCFInterp>(
[…]
