### 2023-04-25 c25364d
WIP on trimming LSQ-based multigrid interpolator
