+++ Source/AmrMesh/CD_EBAMRDataImplem.H
@@ -12,50 +12,81 @@
 #ifndef CD_EBAMRDataImplem_H
 #define CD_EBAMRDataImplem_H
 
+// Chombo includes
+#include <CH_Timer.H>
+
 // Our includes
 #include <CD_EBAMRData.H>
 #include <CD_NamespaceHeader.H>
 
 template <typename T>
-EBAMRData<T>::EBAMRData()
-{}
+EBAMRData<T>::EBAMRData() noexcept
+{
+  CH_TIME("EBAMRData<T>::EBAMRData()");
+  m_reserved = false;
+}
 
 template <typename T>
-EBAMRData<T>::EBAMRData(const int a_size)
+EBAMRData<T>::EBAMRData(const int a_size) noexcept : EBAMRData<T>()
 {
+  CH_TIME("EBAMRData<T>::EBAMRData(int)");
+
   m_data.resize(a_size);
 }
 
 template <typename T>
-EBAMRData<T>::~EBAMRData()
-{}
+EBAMRData<T>::~EBAMRData() noexcept
+{
+  CH_TIME("EBAMRData<T>::~EBAMRData()");
+}
 
 template <typename T>
 RefCountedPtr<LevelData<T>>&
-EBAMRData<T>::operator[](int a_level)
+EBAMRData<T>::operator[](int a_level) noexcept
 {
   return m_data[a_level];
 }
 
 template <typename T>
 const RefCountedPtr<LevelData<T>>&
-EBAMRData<T>::operator[](int a_level) const
+EBAMRData<T>::operator[](int a_level) const noexcept
 {
   return m_data[a_level];
 }
 
 template <typename T>
 void
-EBAMRData<T>::resize(const int a_size)
+EBAMRData<T>::resize(const int a_size) noexcept
 {
   m_data.resize(a_size);
 }
 
 template <typename T>
 int
-EBAMRData<T>::size() const
+EBAMRData<T>::size() const noexcept
 {
   return m_data.size();
+}
+
+template <typename T>
+void
+EBAMRData<T>::reserve() noexcept
+{
+  m_reserved = true;
+}
+
+template <typename T>
+void
+EBAMRData<T>::release() noexcept
+{
+  m_reserved = false;
+}
+
+template <typename T>
+bool
+EBAMRData<T>::isReserved() const noexcept
+{
+  return m_reserved;
 }
 
 template <typename T>
@@ -69,42 +100,42 @@ EBAMRData<T>::exchange() noexcept
 
 template <typename T>
 Vector<RefCountedPtr<LevelData<T>>>&
-EBAMRData<T>::getData()
+EBAMRData<T>::getData() noexcept
 {
   return m_data;
 }
 
 template <typename T>
 const Vector<RefCountedPtr<LevelData<T>>>&
-EBAMRData<T>::getData() const
+EBAMRData<T>::getData() const noexcept
 {
   return m_data;
 }
 
 template <typename T>
 void
-EBAMRData<T>::push_back(RefCountedPtr<LevelData<T>>& a_levelData)
+EBAMRData<T>::push_back(RefCountedPtr<LevelData<T>>& a_levelData) noexcept
 {
   m_data.push_back(a_levelData);
 }
 
 template <typename T>
 const std::string
-EBAMRData<T>::getRealm() const
+EBAMRData<T>::getRealm() const noexcept
 {
   return m_realm;
 }
 
 template <typename T>
 void
-EBAMRData<T>::setRealm(const std::string a_realm)
+EBAMRData<T>::setRealm(const std::string a_realm) noexcept
 {
   m_realm = a_realm;
 }
 
 template <typename T>
 void
[…]
