+++ Source/AmrMesh/CD_EBLeastSquaresMultigridInterpolator.cpp
@@ -152,8 +152,8 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
     m_quadCFInterp->coarseFineInterp(fineAliasOneComp, coarAliasOneComp);
     CH_STOP(t1);
 
-    // Now do the EB. Copy data to scratch data holders. We need the coarse-data to be accessible by the fine data (through the same iterators), so we can't
-    // get away from a copy here (I think).
+    // Now do the EB. Copy data to scratch data holders. We need the coarse-data to be accessible by the
+    // fine data (through the same iterators), so we can't  get away from a copy here (I think).
     CH_START(t2);
     const Interval srcInterv = Interval(icomp, icomp);
     const Interval dstInterv = Interval(m_comp, m_comp);
@@ -168,7 +168,7 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
       const EBCellFAB& srcFine = a_phiFine[dit()];
       const EBCellFAB& srcCoar = m_grownCoarData[dit()];
 
-      // Apply the coarse stencil
+      // Apply the coarse and fine stencils
       constexpr int numComp = 1;
       m_aggCoarStencils[dit()]->apply(dstFine,
                                       srcCoar,
@@ -194,6 +194,7 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterpH(LevelData<EBCellFAB>& a_p
   CH_TIME("EBLeastSquaresMultigridInterpolator::coarseFineInterpH(LD<EBCellFAB>, Interval)");
 
   CH_assert(m_ghostCF * IntVect::Unit <= a_phiFine.ghostVect());
+  CH_assert(a_phiFine.nComp() > a_variables.end());
 
   if (a_phiFine.ghostVect() != m_ghostVector) {
     MayDay::Error("EBLeastSquaresMultigridInterpolator::coarseFineInterp -- number of ghost cells do not match!");
@@ -210,19 +211,27 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterpH(EBCellFAB&       a_phi,
                                                        const Interval   a_variables,
                                                        const DataIndex& a_dit) const noexcept
 {
-  CH_TIME("EBLeastSquaresMultigridInterpolator::coarseFineInterpH(LD<EBCellFAB>, Interval, DataIndex)");
+  CH_TIMERS("EBLeastSquaresMultigridInterpolator::coarseFineInterpH(LD<EBCellFAB>)");
+  CH_TIMER("regular_interp", t1);
+  CH_TIMER("irregular_interp", t2);
+
+  CH_assert(a_phi.nComp() > a_variables.end());
 
   // TLDR: This routine does the coarse-fine interpolation with the coarse-grid data set to zero. This is the kernel version,
   //       operating on a grid patch. It first does a direct kernel for regular data, and then does the interpolation near the
   //       EB after that.
   const Real dxFine = 1.0;
   const Real dxCoar = 1.0 * m_refRat;
+  const Real c1     = 2 * (dxCoar - dxFine) / (dxCoar + dxFine);
+  const Real c2     = -(dxCoar - dxFine) / (dxCoar + 3 * dxFine);
 
   for (int ivar = a_variables.begin(); ivar <= a_variables.end(); ivar++) {
 
     // Do the regular interp on all sides of the current patch.
+    CH_START(t1);
     for (int dir = 0; dir < SpaceDim; dir++) {
       for (SideIterator sit; sit.ok(); ++sit) {
+        const IntVect shift = sign(sit()) * BASISV(dir);
 
         // Regular homogeneous interpolation on side sit() in direction dir
[…]
