+++ Source/AmrMesh/CD_EBLeastSquaresMultigridInterpolator.cpp
@@ -42,8 +42,8 @@ EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator(const E
                                                                          const int          a_weighting) noexcept
 {
   CH_TIMERS("EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator");
-  CH_TIMER("define_regular", t1);
-  CH_TIMER("define_irregular", t2);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::define_regular", t1);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::define_irregular", t2);
 
   CH_assert(a_ghostCF > 0);
   CH_assert(a_refRat % 2 == 0);
@@ -102,9 +102,9 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
                                                       const Interval              a_variables) const noexcept
 {
   CH_TIMERS("EBLeastSquaresMultigridInterpolator::coarseFineInterp");
-  CH_TIMER("regular_interp", t1);
-  CH_TIMER("copyTo", t2);
-  CH_TIMER("irregular_interp", t3);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::regular_interp", t1);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::copyTo", t2);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::irregular_interp", t3);
 
   CH_assert(m_ghostCF * IntVect::Unit <= a_phiFine.ghostVect());
   CH_assert(a_phiFine.nComp() > a_variables.end());
@@ -156,13 +156,13 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
                                       m_comp,
                                       icomp,
                                       numComp,
-                                      false); // true/false => increment/not increment
+                                      false);
       m_aggFineStencils[dit()]->apply(dstFine,
                                       srcFine,
                                       icomp,
                                       icomp,
                                       numComp,
-                                      true); // true/false => increment/not increment
+                                      true); 
     }
     CH_STOP(t3);
   }
@@ -193,8 +193,8 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterpH(EBCellFAB&       a_phi,
                                                        const DataIndex& a_dit) const noexcept
 {
   CH_TIMERS("EBLeastSquaresMultigridInterpolator::coarseFineInterpH(LD<EBCellFAB>)");
-  CH_TIMER("regular_interp", t1);
-  CH_TIMER("irregular_interp", t2);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::regular_interp", t1);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::irregular_interp", t2);
 
   CH_assert(a_phi.nComp() > a_variables.end());
 
@@ -304,7 +304,7 @@ EBLeastSquaresMultigridInterpolator::defineGhostRegions() noexcept
 
       // 2. Only include ghost cells that are within range m_ghostCF of an irregular grid cell
       IntVectSet irreg = ebisbox.getIrregIVS(cellBox);
-      irreg.grow(2 * m_refRat);
+      irreg.grow(m_ghostCF);
       m_ghostCells[dit()] &= irreg;
     }
   }
@@ -332,11 +332,11 @@ void
[…]
