### 2023-04-26 88fc17b
WIP towards not actually storing the stencils
