+++ Source/AmrMesh/CD_CoarseInterpQuadCF.cpp
@@ -1,0 +1,382 @@
+/* chombo-discharge
+ * Copyright © 2023 SINTEF Energy Research.
+ * Please refer to Copyright.txt and LICENSE in the chombo-discharge root directory.
+ */
+
+/*!
+  @file   CD_CoarseInterpQuadCF.cpp
+  @brief  Implementation of CD_CoarseInterpQuadCF.H
+  @author Robert Marskar
+*/
+
+// Chombo includes
+#include <CH_Timer.H>
+#include <DenseIntVectSet.H>
+#include <NeighborIterator.H>
+
+// Our includes
+#include <CD_CoarseInterpQuadCF.H>
+#include <CD_NamespaceHeader.H>
+
+CoarseInterpQuadCF::CoarseInterpQuadCF() noexcept
+{
+  CH_TIME("CoarseInterpQuadCF::CoarseInterpQuadCF()");
+
+  m_isDefined = false;
+}
+
+CoarseInterpQuadCF::~CoarseInterpQuadCF() noexcept { CH_TIME("CoarseInterpQuadCF::~CoarseInterpQuadCF()"); }
+
+void
+CoarseInterpQuadCF::define(const DisjointBoxLayout& a_dblFine,
+                           const ProblemDomain&     a_domainCoar,
+                           const DataIndex&         a_dit,
+                           const Box&               a_fineGhostCells,
+                           const int                a_refRat,
+                           const int                a_ignoreDir) noexcept
+{
+  CH_TIME("CoarseInterpQuadCF::define");
+
+  m_dblFine    = a_dblFine;
+  m_domainCoar = a_domainCoar;
+  m_dit        = a_dit;
+  m_refRat     = a_refRat;
+  m_ignoreDir  = a_ignoreDir;
+  m_stencilBox = coarsen(a_fineGhostCells, m_refRat);
+
+  // Define stencil objects.
+  for (int dir = 0; dir < SpaceDim; dir++) {
+    if (dir != a_ignoreDir) {
+      m_whichFirstDerivStencils[dir].define(m_stencilBox, 1);
+      m_firstDerivStencils[dir].define(m_stencilBox, 1);
+      m_secondDerivStencils[dir].define(m_stencilBox, 1);
+    }
+  }
+#if CH_SPACEDIM == 3
+  m_mixedDerivStencils.define(m_stencilBox, 1);
+#endif
+
+  // Compute stencils.
+  this->defineFirstDerivStencils();
+  this->defineSecondDerivStencils();
+  this->defineMixedDerivStencils();
+
+  m_isDefined = true;
+}
+
+void
+CoarseInterpQuadCF::defineFirstDerivStencils() noexcept
+{
+  CH_TIME("CoarseInterpQuadCF::defineFirstDerivStencils");
+
+  // Stencils will have a radius of 2, so figure out which cells we are allowed to use.
+  DenseIntVectSet validCells(grow(m_stencilBox, 2), true);
+  validCells -= m_dblFine[m_dit];
+  NeighborIterator nit(m_dblFine);
+  for (nit.begin(m_dit); nit.ok(); ++nit) {
+    validCells -= m_dblFine[nit()];
+  }
+  validCells &= m_domainCoar;
+
+  // Go through the cells and compute finite difference approximations.
+  for (int dir = 0; dir < SpaceDim; dir++) {
+    if (dir != m_ignoreDir) {
+      BaseFab<DerivStencil>& derivStencilsDir = m_firstDerivStencils[dir];
+
+      for (BoxIterator bit(m_stencilBox); bit.ok(); ++bit) {
+        const IntVect ivCoar = bit();
+
+        DerivStencil& derivSten = derivStencilsDir(ivCoar, 0);
+        derivSten.clear();
+
[…]
