+++ Source/AmrMesh/CD_CoarseInterpQuadCF.H
@@ -1,0 +1,192 @@
+/* chombo-discharge
+ * Copyright © 2023 SINTEF Energy Research.
+ * Please refer to Copyright.txt and LICENSE in the chombo-discharge root directory.
+ */
+
+/*!
+  @file   CD_CoarseInterpQuadCF.H
+  @brief  Supporting class for doing the coarse-side interpolation for multigrid
+  @author Robert Marskar
+*/
+
+#ifndef CD_CoarseInterpQuadCF_H
+#define CD_CoarseInterpQuadCF_H
+
+// Chombo includes
+#include <IntVect.H>
+#include <FArrayBox.H>
+#include <LoHiSide.H>
+#include <DerivStencil.H>
+#include <DisjointBoxLayout.H>
+
+// Our includes
+#include <CD_NamespaceHeader.H>
+
+/*!
+  @brief Class for computing the required Taylor terms for doing the orthogonal extrapolation to the ghost cell position. This
+  is used in conjuction with EBLeastSquaresMultigridInterpolator.H, and should be understood in that context only. 
+
+  @details This should be defined over the fine grid. It will store stencils for computing the various required terms in the 
+  Taylor series, and the user can call various functions for getting those terms. 
+
+  @todo This class stores all stencils explicitly. We might want to consider a "default case" (e.g., centered differencing) for compressing memory.
+*/
+class CoarseInterpQuadCF
+{
+public:
+  /*!
+    @brief Default constructor. Must subsequently call the define function
+  */
+  CoarseInterpQuadCF() noexcept;
+
+  /*!
+    @brief Destructor.
+  */
+  virtual ~CoarseInterpQuadCF() noexcept;
+
+  /*!
+    @brief Define function. Puts object in usable state.
+    @param[in] a_dblFine Fine grids
+    @param[in] a_domainCoar Coarse domain
+    @param[in] a_dit Data index on the fine grid
+    @param[in] a_fineGhostCells Ghost cells to interpolate (on the fine level)
+    @param[in] a_refRat Refinement factor between coarse and fine
+    @param[in] a_ignoreDir Coordinate to ignore when doing the interpolation. I.e., we interpolate 
+    in all coordinates except this one.
+  */
+  virtual void
+  define(const DisjointBoxLayout& a_dblFine,
+         const ProblemDomain&     a_domainCoar,
+         const DataIndex&         a_dit,
+         const Box&               a_fineGhostCells,
+         const int                a_refRat,
+         const int                a_ignoreDir) noexcept;
+
+  /*!
+    @brief Compute first-derivative Taylor term in direction a_dir. 
+    @param[in] a_coarPhi Coarse-side phi. 
+    @param[in] a_ivCoar Cell where we compute.
+    @param[in] a_dir Direction of derivative. Must be different from m_ignoreDir
+    @param[in] a_coarVar a_dir Direction of derivative. Must be different from m_ignoreDir
+  */
+  virtual Real
+  computeFirstDeriv(const FArrayBox& a_coarPhi,
+                    const IntVect&   a_ivCoar,
+                    const int        a_dir,
+                    const int        a_coarVar) const noexcept;
+
+  /*!
+    @brief Compute second-derivative Taylor term in direction a_dir. 
+    @param[in] a_coarPhi Coarse-side phi. 
+    @param[in] a_ivCoar Cell where we compute.
+    @param[in] a_dir Direction of derivative. Must be different from m_ignoreDir
+    @param[in] a_coarVar a_dir Direction of derivative. Must be different from m_ignoreDir
[…]
