### 2023-04-26 b2e6df7
Not storing first deriv stencil explicitly
