### 2023-04-26 c27d8b6
Compressing. Need to get rid of explicit mixed stencils now
