+++ Source/AmrMesh/CD_CoarseInterpQuadCF.cpp
@@ -56,17 +56,16 @@ CoarseInterpQuadCF::define(const DisjointBoxLayout& a_dblFine,
 #endif
 
   // Compute stencils.
-  this->defineFirstDerivStencils();
-  this->defineSecondDerivStencils();
+  this->defineStencils();
   this->defineMixedDerivStencils();
 
   m_isDefined = true;
 }
 
 void
-CoarseInterpQuadCF::defineFirstDerivStencils() noexcept
+CoarseInterpQuadCF::defineStencils() noexcept
 {
-  CH_TIME("CoarseInterpQuadCF::defineFirstDerivStencils");
+  CH_TIME("CoarseInterpQuadCF::defineStencils");
 
   // Stencils will have a radius of 2, so figure out which cells we are allowed to use.
   DenseIntVectSet validCells(grow(m_stencilBox, 2), true);
@@ -96,79 +95,22 @@ CoarseInterpQuadCF::defineFirstDerivStencils() noexcept
         // Switch between differencing formulas. We want to use second order differencing if we can, but use
         // first order if we must.
         if (useLo && useHi) {
-          m_firstDerivStencils[dir](ivCoar, 0) = FirstDerivStencil::Centered2;
+          m_firstDerivStencils[dir](ivCoar, 0)  = FirstDerivStencil::Centered2;
+          m_secondDerivStencils[dir](ivCoar, 0) = SecondDerivStencil::Centered2;
         }
         else if (useLo && useLoLo) {
-          m_firstDerivStencils[dir](ivCoar, 0) = FirstDerivStencil::Backward2;
+          m_firstDerivStencils[dir](ivCoar, 0)  = FirstDerivStencil::Backward2;
+          m_secondDerivStencils[dir](ivCoar, 0) = SecondDerivStencil::Backward1;
         }
         else if (useHi && useHiHi) {
-          m_firstDerivStencils[dir](ivCoar, 0) = FirstDerivStencil::Forward2;
+          m_firstDerivStencils[dir](ivCoar, 0)  = FirstDerivStencil::Forward2;
+          m_secondDerivStencils[dir](ivCoar, 0) = SecondDerivStencil::Forward1;
         }
         else if (useLo && !useHi) {
           m_firstDerivStencils[dir](ivCoar, 0) = FirstDerivStencil::Backward1;
         }
         else if (!useLo && useHi) {
           m_firstDerivStencils[dir](ivCoar, 0) = FirstDerivStencil::Forward1;
-        }
-      }
-    }
-  }
-}
-
-void
-CoarseInterpQuadCF::defineSecondDerivStencils() noexcept
-{
-  CH_TIME("CoarseInterpQuadCF::defineSecondDerivStencils");
-
-  // Stencils will have a radius of 2, so figure out which cells we are allowed to use.
-  DenseIntVectSet validCells(grow(m_stencilBox, 2), true);
-  validCells -= m_dblFine[m_dit];
-  NeighborIterator nit(m_dblFine);
-  for (nit.begin(m_dit); nit.ok(); ++nit) {
-    validCells -= m_dblFine[nit()];
-  }
-  validCells &= m_domainCoar;
-
-  // Go through the cells and compute finite difference approximations.
-  for (int dir = 0; dir < SpaceDim; dir++) {
-    if (dir != m_ignoreDir) {
-      BaseFab<DerivStencil>& derivStencilsDir = m_secondDerivStencils[dir];
-
-      for (BoxIterator bit(m_stencilBox); bit.ok(); ++bit) {
-        const IntVect ivCoar = bit();
-
[…]
