### 2023-04-26 8cd83fb
More memory trim
