+++ Source/AmrMesh/CD_CoarseInterpQuadCF.cpp
@@ -44,6 +44,20 @@ CoarseInterpQuadCF::define(const DisjointBoxLayout& a_dblFine,
   m_ignoreDir  = a_ignoreDir;
   m_stencilBox = coarsen(a_fineGhostCells, m_refRat);
 
+  // Store the coordinates tangential to ignoreDir. Needed for mix-deriv stencil.
+  if (m_ignoreDir == 0) {
+    m_tanDir1 = 1;
+    m_tanDir2 = 2;
+  }
+  else if (m_ignoreDir == 1) {
+    m_tanDir1 = 0;
+    m_tanDir2 = 2;
+  }
+  else if (m_ignoreDir == 2) {
+    m_tanDir1 = 0;
+    m_tanDir2 = 1;
+  }
+
   // Define stencil objects.
   for (int dir = 0; dir < SpaceDim; dir++) {
     if (dir != a_ignoreDir) {
@@ -53,6 +67,7 @@ CoarseInterpQuadCF::define(const DisjointBoxLayout& a_dblFine,
   }
 #if CH_SPACEDIM == 3
   m_mixedDerivStencils.define(m_stencilBox, 1);
+  m_explicitMixedDerivStencils.define(m_stencilBox, 1);
 #endif
 
   // Compute stencils.
@@ -132,55 +147,55 @@ CoarseInterpQuadCF::defineMixedDerivStencils() noexcept
   }
   validCells &= m_domainCoar;
 
+  constexpr Real factor = 1. / 16.;
+
   // Go through the cells and compute finite difference approximations to the mixed derivative. We do this ala Chombo
   // and average the mixed-derivative stencils on edges of the cell since some of the cells we otherwise would need
   // might be covered by the fine grid.
   for (BoxIterator bit(m_stencilBox); bit.ok(); ++bit) {
     const IntVect ivCoar = bit();
 
-    DerivStencil& derivSten = m_mixedDerivStencils(ivCoar, 0);
-    derivSten.clear();
-
-    int dir1 = -1;
-    int dir2 = -1;
-
-    if (m_ignoreDir == 0) {
-      dir1 = 1;
-      dir2 = 2;
+    if (validCells.contains(grow(Box(ivCoar, ivCoar), 1))) {
+      m_mixedDerivStencils(ivCoar, 0) = MixedDerivStencil::Standard;
     }
-    else if (m_ignoreDir == 1) {
-      dir1 = 0;
-      dir2 = 2;
-    }
-    else if (m_ignoreDir == 2) {
-      dir1 = 0;
-      dir2 = 1;
-    }
+    else {
+      m_mixedDerivStencils(ivCoar, 0) = MixedDerivStencil::Explicit;
 
-    // Quadrant which may or may not be available.
-    Vector<Box> quadrants;
+      DerivStencil& derivSten = m_explicitMixedDerivStencils(ivCoar, 0);
+      derivSten.clear();
 
-    quadrants.push_back(Box(ivCoar, ivCoar + BASISV(dir1) + BASISV(dir2)));
-    quadrants.push_back(Box(ivCoar - BASISV(dir1), ivCoar + BASISV(dir2)));
-    quadrants.push_back(Box(ivCoar - BASISV(dir1) - BASISV(dir2), ivCoar));
-    quadrants.push_back(Box(ivCoar - BASISV(dir2), ivCoar + BASISV(dir1)));
+      // Quadrant which may or may not be available.
+      int numQuadrants = 0;
 
-    int numQuadrants;
+      const Box baseQuadrant(ivCoar, ivCoar + BASISV(m_tanDir1) + BASISV(m_tanDir2));
 
[…]
