### 2023-04-26 7f6e2eb
Final updates on interpolator. Getting rid of QuadCFInterp now
