### 2023-04-26 0f52ddc
Add some memory trackers to see where our memory goes. These will be removed later
