### 2023-04-26 0245297
Bit more transparent where the memory is going
