+++ Source/Driver/CD_Driver.cpp
@@ -1481,37 +1481,40 @@ Driver::setupFresh(const int a_initialRegrids)
   this->allocateInternals();
 
   // Provide TimeStepper with geometry in case it needs it.
-  pout() << "before set comp geom" << endl;
+  pout() << "before/after set comp geom" << endl;
   MemoryReport::getMaxMinMemoryUsage();
   m_timeStepper->setComputationalGeometry(m_computationalGeometry);
+  MemoryReport::getMaxMinMemoryUsage();
+  pout() << endl;
 
   // TimeStepper setup. This instantiatse solvers (but does not necessarily fill them with data).
-  pout() << "before setup solvers" << endl;
-  MemoryReport::getMaxMinMemoryUsage();    
+  pout() << "before/after setup solvers" << endl;
+  MemoryReport::getMaxMinMemoryUsage();
   m_timeStepper->setupSolvers();
   m_timeStepper->synchronizeSolverTimes(m_timeStep, m_time, m_dt);
+  MemoryReport::getMaxMinMemoryUsage();
+  pout() << endl;
 
   // Set up the AMR operators
   m_timeStepper->registerOperators();
-  pout() << "before regrid opers" << endl;
-  MemoryReport::getMaxMinMemoryUsage();    
+  pout() << "before/after regrid opers" << endl;
+  MemoryReport::getMaxMinMemoryUsage();
   m_amr->regridOperators(lmin);
-  pout() << "after regrid opers" << endl;
-  MemoryReport::getMaxMinMemoryUsage();      
-
+  MemoryReport::getMaxMinMemoryUsage();
+  pout() << endl;
 
   // Fill solves with initial data
-  pout() << "before ts alloc" << endl;
-  MemoryReport::getMaxMinMemoryUsage();    
+  pout() << "before/after timestepper alloc" << endl;
+  MemoryReport::getMaxMinMemoryUsage();
   m_timeStepper->allocate();
-  pout() << "after ts alloc" << endl;  
-  pout() << "before init data" << endl;
-  MemoryReport::getMaxMinMemoryUsage();    
-  m_timeStepper->initialData();
-  pout() << "after init data" << endl;
-  MemoryReport::getMaxMinMemoryUsage();      
+  MemoryReport::getMaxMinMemoryUsage();
+  pout() << endl;
 
-  
+  pout() << "before/after init data" << endl;
+  MemoryReport::getMaxMinMemoryUsage();
+  m_timeStepper->initialData();
+  MemoryReport::getMaxMinMemoryUsage();
+  pout() << endl;
 
   // If called for -- we can perform a
   if (m_doInitLoadBalancing) {
