### 2023-04-27 2d50368
Added explicit error messages for multigrid
