+++ Source/Elliptic/CD_EBHelmholtzRobinEBBC.cpp
@@ -224,7 +224,14 @@ EBHelmholtzRobinEBBC::define()
           fluxStencil.clear();
         }
       }
-      else { // Dead cell
+      else {
+        // Dead cell. No flux
+        const std::string baseErr = "EBHelmholtzRobinEBBC::define - dead cell on domain = ";
+        const std::string vofErr  = " on vof = ";
+        const std::string impErr  = " (this may cause multigrid divergence)";
+
+        std::cout << baseErr << m_eblg.getDomain() << vofErr << vof << impErr << std::endl;
+
         fluxStencil.clear();
       }
     };
