+++ Source/Elliptic/CD_MFHelmholtzRobinEBBC.cpp
@@ -194,7 +194,14 @@ MFHelmholtzRobinEBBC::defineSinglePhase()
           fluxStencil.clear();
         }
       }
-      else { // Dead cell
+      else {
+        // Dead cell. No flux.
+        const std::string baseErr = "MFHelmholtzRobinEBBC::defineSinglePhase - dead cell on domain = ";
+        const std::string vofErr  = " on vof = ";
+        const std::string impErr  = " (this may cause multigrid divergence)";
+
+        std::cout << baseErr << m_eblg.getDomain() << vofErr << vof << impErr << std::endl;
+
         fluxStencil.clear();
       }
     };
