+++ Source/Electrostatics/CD_MFHelmholtzElectrostaticEBBC.cpp
@@ -141,7 +141,13 @@ MFHelmholtzElectrostaticEBBC::defineSinglePhase()
         stencils(vof, m_comp) *= areaFrac / m_dx;
       }
       else {
-        // Dead cell. No flux.
+        // Dead cell. No flux
+        const std::string baseErr = "MFHelmholtzElectrostaticEBBC::defineSinglePhase - dead cell on domain = ";
+        const std::string vofErr  = " on vof = ";
+        const std::string impErr  = " (this may cause multigrid divergence)";
+
+        std::cout << baseErr << m_eblg.getDomain() << vofErr << vof << impErr << std::endl;
+
         weights(vof, m_comp) = 0.0;
         stencils(vof, m_comp).clear();
       }
