+++ Source/Elliptic/CD_EBHelmholtzDirichletEBBC.cpp
@@ -172,6 +172,12 @@ EBHelmholtzDirichletEBBC::define()
       }
       else {
         // Dead cell. No flux.
+        const std::string baseErr = "EBHelmholtzDirichletEBBC::define - dead cell on domain = ";
+        const std::string vofErr  = " on vof = ";
+        const std::string impErr  = " (this may cause multigrid divergence)";
+
+        std::cout << baseErr << m_eblg.getDomain() << vofErr << vof << impErr << std::endl;
+
         weights(vof, m_comp) = 0.0;
         stencils(vof, m_comp).clear();
       }
