+++ Physics/ItoKMC/PlasmaModels/ItoKMCAir3LFA/CD_ItoKMCAir3LFA.cpp
@@ -101,11 +101,11 @@ ItoKMCAir3LFA::ItoKMCAir3LFA()
   auto y1 = std::make_shared<ItoKMCPhotoReaction>(0, std::list<size_t>{0, 1});
   m_photoReactions.emplace_back(y1);
 
+  // Read input data
+  this->readTables();  
+
   // Run the rest of the define method.
   this->define();
-
-  // Read input data
-  this->readTables();
 }
 
 ItoKMCAir3LFA::~ItoKMCAir3LFA() { CH_TIME("ItoKMCAir3LFA::~ItoKMCAir3LFA"); }
