### 2023-04-27 e2abd96
Using transient buffer for memory compression in lsq interpolator
