+++ Source/AmrMesh/CD_EBLeastSquaresMultigridInterpolator.cpp
@@ -90,6 +90,7 @@ EBLeastSquaresMultigridInterpolator::EBLeastSquaresMultigridInterpolator(const E
   this->defineBuffers();
   this->defineCoarseInterp();
   this->defineStencilsEBCF();
+  this->undefineBuffers();
   CH_STOP(t2);
 }
 
@@ -110,9 +111,11 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
                                                       const Interval              a_variables) const noexcept
 {
   CH_TIMERS("EBLeastSquaresMultigridInterpolator::coarseFineInterp");
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::buffer_define", t0);
   CH_TIMER("EBLeastSquaresMultigridInterpolator::copy_exchange", t1);
   CH_TIMER("EBLeastSquaresMultigridInterpolator::regular_interp", t2);
   CH_TIMER("EBLeastSquaresMultigridInterpolator::irregular_interp", t3);
+  CH_TIMER("EBLeastSquaresMultigridInterpolator::buffer_undefine", t4);
 
   CH_assert(m_ghostCF * IntVect::Unit <= a_phiFine.ghostVect());
   CH_assert(a_phiFine.nComp() > a_variables.end());
@@ -122,6 +125,10 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
     MayDay::Error("EBLeastSquaresMultigridInterpolator::coarseFineInterp -- number of ghost cells do not match!");
   }
 
+  CH_START(t0);
+  this->defineBuffers();
+  CH_STOP(t0);
+
   // Interpolate all variables near the EB. We will copy a_phiCoar to m_grownCoarData which holds the data on the coarse grid cells around each fine-grid
   // patch. Note that m_grownCoarData provides a LOCAL view of the coarse grid around each fine-level patch, so we can apply the stencils directly.
   for (int icomp = a_variables.begin(); icomp <= a_variables.end(); icomp++) {
@@ -130,8 +137,8 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
     CH_START(t1);
     const Interval srcInterv = Interval(icomp, icomp);
     const Interval dstInterv = Interval(m_comp, m_comp);
-    a_phiCoar.copyTo(srcInterv, m_grownCoarData, dstInterv);
-    m_grownCoarData.exchange();
+    a_phiCoar.copyTo(srcInterv, *m_grownCoarData, dstInterv);
+    m_grownCoarData->exchange();
     CH_STOP(t1);
 
     // Do the regular interpolation as if the EB is not there.
@@ -142,7 +149,7 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
     LevelData<FArrayBox> coarAliasOneComp;
 
     aliasEB(fineAlias, (LevelData<EBCellFAB>&)a_phiFine);
-    aliasEB(coarAlias, (LevelData<EBCellFAB>&)m_grownCoarData);
+    aliasEB(coarAlias, (LevelData<EBCellFAB>&)*m_grownCoarData);
 
     aliasLevelData(fineAliasOneComp, &fineAlias, Interval(icomp, icomp));
     aliasLevelData(coarAliasOneComp, &coarAlias, Interval(0, 0));
@@ -163,7 +170,7 @@ EBLeastSquaresMultigridInterpolator::coarseFineInterp(LevelData<EBCellFAB>&     
     for (DataIterator dit = a_phiFine.dataIterator(); dit.ok(); ++dit) {
[…]
