+++ Source/AmrMesh/CD_PhaseRealm.cpp
@@ -169,14 +169,6 @@ PhaseRealm::regridOperators(const int a_lmin)
 
     Timer timer("PhaseRealm::regridOperators(int)");
 
-    pout() << "before/after buffer define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
-    timer.startEvent("Define buffers");
-    this->defineBuffers(a_lmin);
-    timer.stopEvent("Define buffers");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
-
     pout() << "before/after coarave define" << endl;
     MemoryReport::getMaxMinMemoryUsage();
     timer.startEvent("EBCoarAve");
@@ -494,48 +486,6 @@ PhaseRealm::defineLevelSet(const int a_lmin, const int a_numGhost)
           fab.setVal(minVal, comp);
         }
       }
-    }
-  }
-}
-
-void
-PhaseRealm::defineBuffers(const int a_lmin) noexcept
-{
-  CH_TIME("PhaseRealm::defineBuffers");
-  if (m_verbose) {
-    pout() << "PhaseRealm::defineBuffers" << endl;
-  }
-
-  m_coFiBufferCell.resize(1 + m_finestLevel);
-  m_coFiBufferFlux.resize(1 + m_finestLevel);
-  m_coFiBufferEB.resize(1 + m_finestLevel);
-
-  for (int lvl = a_lmin; lvl <= m_finestLevel; lvl++) {
-    const bool hasFine = lvl < m_finestLevel;
-    const bool hasCoar = lvl > 0;
-
-    if (hasFine) {
-      CH_assert(!m_eblgCoFi[lvl].isNull());
-
-      const DisjointBoxLayout& coFiDBL   = m_eblgCoFi[lvl]->getDBL();
-      const EBISLayout&        coFiEBISL = m_eblgCoFi[lvl]->getEBISL();
-
-      // Define coarsened fine data.
-      LayoutData<IntVectSet> irregSetsCoFi(coFiDBL);
-      for (DataIterator dit(coFiDBL); dit.ok(); ++dit) {
-        irregSetsCoFi[dit()] = coFiEBISL[dit()].getIrregIVS(coFiDBL[dit()]);
-      }
-
-      BaseIVFactory<Real> ebFactory(coFiEBISL, irregSetsCoFi);
-
-      m_coFiBufferCell[lvl] = RefCountedPtr<LevelData<EBCellFAB>>(
-        new LevelData<EBCellFAB>(coFiDBL, 1, IntVect::Zero, EBCellFactory(coFiEBISL)));
-
-      m_coFiBufferFlux[lvl] = RefCountedPtr<LevelData<EBFluxFAB>>(
-        new LevelData<EBFluxFAB>(coFiDBL, 1, IntVect::Zero, EBFluxFactory(coFiEBISL)));
-
-      m_coFiBufferEB[lvl] = RefCountedPtr<LevelData<BaseIVFAB<Real>>>(
-        new LevelData<BaseIVFAB<Real>>(coFiDBL, 1, IntVect::Zero, BaseIVFactory<Real>(coFiEBISL, irregSetsCoFi)));
     }
   }
 }
@@ -1205,24 +1155,6 @@ PhaseRealm::getLevelset() const
   }
 
   return m_levelset;
-}
-
-EBAMRCellData&
-PhaseRealm::getCoFiBufferCell() const noexcept
-{
-  return this->m_coFiBufferCell;
-}
-
-EBAMRFluxData&
-PhaseRealm::getCoFiBufferFlux() const noexcept
-{
-  return this->m_coFiBufferFlux;
-}
-
-EBAMRIVData&
[…]
