+++ Source/AmrMesh/CD_PhaseRealm.cpp
@@ -127,29 +127,41 @@ PhaseRealm::regridBase(const int a_lmin)
 
     Timer timer("PhaseRealm::regridBase(int)");
 
-    pout() << "before/after levelgrid define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after levelgrid define" << endl;
+      MemoryReport::getMaxMinMemoryUsage();
+    }
     timer.startEvent("Define EBLevelGrid");
     this->defineEBLevelGrid(a_lmin);
     timer.stopEvent("Define EBLevelGrid");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
+    if (m_profile) {
+      MemoryReport::getMaxMinMemoryUsage();
+      pout() << endl;
+    }
 
-    pout() << "before/after neighbor define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after neighbor define" << endl;
+      MemoryReport::getMaxMinMemoryUsage();
+    }
     timer.startEvent("Define neighbors");
     this->defineNeighbors(a_lmin);
     timer.stopEvent("Define neighbors");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
+    if (m_profile) {
+      MemoryReport::getMaxMinMemoryUsage();
+      pout() << endl;
+    }
 
-    pout() << "before/after vofit define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after vofit define" << endl;
+      MemoryReport::getMaxMinMemoryUsage();
+    }
     timer.startEvent("Define VoFIterators");
     this->defineVofIterator(a_lmin);
     timer.stopEvent("Define VoFIterators");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
+    if (m_profile) {
+      MemoryReport::getMaxMinMemoryUsage();
+      pout() << endl;
+    }
 
     if (m_profile) {
       timer.eventReport(pout());
@@ -169,117 +181,173 @@ PhaseRealm::regridOperators(const int a_lmin)
 
     Timer timer("PhaseRealm::regridOperators(int)");
 
-    pout() << "before/after coarave define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after coarave define" << endl;
+      MemoryReport::getMaxMinMemoryUsage();
+    }
     timer.startEvent("EBCoarAve");
     this->defineEBCoarAve(a_lmin);
     timer.stopEvent("EBCoarAve");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
+    if (m_profile) {
+      MemoryReport::getMaxMinMemoryUsage();
+      pout() << endl;
+    }
 
-    pout() << "before/after mg define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after mg define" << endl;
+      MemoryReport::getMaxMinMemoryUsage();
+    }
     timer.startEvent("Multigrid interpolator");
     this->defineEBMultigrid(a_lmin);
     timer.stopEvent("Multigrid interpolator");
-    MemoryReport::getMaxMinMemoryUsage();
-    pout() << endl;
+    if (m_profile) {
+      MemoryReport::getMaxMinMemoryUsage();
+      pout() << endl;
+    }
 
-    pout() << "before/after fillpatch define" << endl;
-    MemoryReport::getMaxMinMemoryUsage();
+    if (m_profile) {
+      pout() << "before/after fillpatch define" << endl;
[…]

[…]
