### 2023-04-27 8ca362d
added writeDataReportTransient() - not finished
