+++ Physics/StreamerInception/CD_StreamerInceptionStepperImplem.H
@@ -3391,6 +3391,126 @@ StreamerInceptionStepper<P, F, C>::writeReportTransient() const noexcept
 #endif
 }
 
+template<typename P, typename F, typename C>
+void
+StreamerInceptionStepper<P,F,C>::writeDataReportTransient(const Real& a_voltage) const noexcept
+{
+  CH_TIME("StreamerInceptionStepper::writeDataReportTransient");
+  if(m_verbosity > 5){
+    pout() << "StreamerInceptionStepper::writeDataReportTransient" << endl;
+  }
+
+  std::vector<RealVect> pos;
+  std::vector<Real> E;
+  std::vector<Real> dA;
+  std::vector<Real> dV;
+  std::vector<Real> ionDensityVector;
+
+  const RealVect probLo = m_amr->getProbLo();
+  
+  for(int lvl = 0; lvl <= m_amr->getFinestLevel(); ++lvl){
+    const DisjointBoxLayout& dbl = m_amr->getGrids(m_realm)[lvl];
+    const EBISLayout& ebisl = m_amr->getEBISLayout(m_realm, m_phase)[lvl];
+
+    const LevelData<BaseFab<bool>>& validCellsLD = *m_amr->getValidCells(m_realm)[lvl];
+
+    const Real dx = m_amr->getDx()[lvl];
+    const Real vol = std::pow(dx, SpaceDim);
+    const Real area = std::pow(dx, SpaceDim -1);
+    
+    for(DataIterator dit = dbl.dataIterator(); dit.ok(); ++dit){
+      const EBISBox& ebisbox = ebisl[dit()];
+      const BaseFab<bool>& validCells = validCellsLD[dit()];
+
+      const EBCellFAB& inceptionIntegral = (*m_inceptionIntegral[lvl])[dit()];
+      
+      const EBCellFAB& electricField = (*m_electricFieldPlus[lvl])[dit()];
+      const FArrayBox& electricFieldReg = electricField.getFArrayBox();
+
+      const EBCellFAB& ionDensity = (*m_ionSolver->getPhi()[lvl])[dit()];
+      const FArrayBox& ionDensityReg = ionDensity.getFArrayBox();
+      
+      auto regularKernel = [&](const IntVect& iv) -> void{
+	if(validCells(iv, 0) && ebisbox.isRegular(iv)){
+	  pos.push_back(probLo + (0.5*RealVect::Unit + RealVect(iv))*dx);
+	  dA.push_back(0.0);
+	
+	  const RealVect EE = RealVect(D_DECL(electricFieldReg(iv, 0), electricFieldReg(iv,1), electricFieldReg(iv, 2)));
+	  const Real currE = std::abs(a_voltage*EE.vectorLength());
+	  const Real alpha = m_alpha(currE);
+	  const Real eta = m_eta(currE);
+	
+	  E.push_back(currE);
+	  ionDensityVector.push_back(ionDensityReg(iv, 0));
+
+	  if(alpha >= eta){
+	    dV.push_back(vol);
+	  }
+	  else{
+	    dV.push_back(0.0);
+	  }
+	}
+      };
+      auto irregularKernel = [&](const VolIndex& vof) -> void{
+	if(validCells(vof.gridIndex())){
+	  pos.push_back(probLo + Location::position(Location::Cell::Centroid, vof, ebisbox, dx));
+	
+	  const RealVect EE = RealVect(D_DECL(electricField(vof, 0), electricField(vof,1), electricField(vof, 2)));
+	  const Real currE = std::abs(a_voltage*EE.vectorLength());
+	  const Real alpha = m_alpha(currE);
+	  const Real eta = m_eta(currE);
+	
+	  E.push_back(currE);
+	  ionDensityVector.push_back(ionDensity(vof, 0));
+	  
[…]
