+++ Source/AmrMesh/CD_EBFineInterp.cpp
@@ -142,6 +142,89 @@ EBFineInterp::defineWeights() noexcept
 }
 
 void
+EBFineInterp::interpolate(LevelData<EBCellFAB>&       a_fineData,
+                          const LevelData<EBCellFAB>& a_coarData,
+                          const Interval&             a_variables,
+                          const EBFineInterp::Type&   a_interpType) const noexcept
+{
+  CH_TIMERS("EBFineInterp::interpolate(LD<EBCellFAB>)");
+  CH_TIMER("EBFineInterp::define_buffer", t1);
+  CH_TIMER("EBFineInterp::copyTo", t2);
+
+  CH_assert(m_isDefined);
+  CH_assert(a_fineData.nComp() > a_variables.end());
+  CH_assert(a_coarData.nComp() > a_variables.end());
+
+  CH_START(t1);
+  const DisjointBoxLayout& dblCoFi   = m_eblgCoFi.getDBL();
+  const EBISLayout&        ebislCoFi = m_eblgCoFi.getEBISL();
+  LevelData<EBCellFAB>     coarsenedFineData(dblCoFi, 1, IntVect::Zero, EBCellFactory(ebislCoFi));
+  CH_STOP(t1);
+
+  for (int ivar = a_variables.begin(); ivar <= a_variables.end(); ivar++) {
+    CH_START(t2);
+    a_coarData.copyTo(Interval(ivar, ivar), coarsenedFineData, Interval(0, 0));
+    CH_STOP(t2);
+
+    for (DataIterator dit(dblCoFi); dit.ok(); ++dit) {
+      switch (a_interpType) {
+      case EBFineInterp::Type::Arithmetic: {
+        this->regridArithmetic(a_fineData, a_coarData, a_variables);
+
+        break;
+      }
+      case EBFineInterp::Type::ConservativeNoSlopes: {
+        this->regridConservativeNoSlopes(a_fineData, a_coarData, a_variables);
+
+        break;
+      }
+      case EBFineInterp::Type::ConservativeMinMod: {
+        //        this->regridMinMod(a_fineData, a_coarData, a_variables);
+        MayDay::Error("EBFineInterp::interpolate - logic bust");
+
+        break;
+      }
+      default: {
+        MayDay::Error("EBFineInterp::interpolate - logic bust. Interpolation type not supported");
+
+        break;
+      }
+      }
+    }
+  }
+}
+
+void
+EBFineInterp::regridArithmetic(LevelData<EBCellFAB>&       a_fineData,
+                               const LevelData<EBCellFAB>& a_coarData,
+                               const Interval&             a_variables) const noexcept
+{
+  CH_TIMERS("EBFineInterp::regridArithmetic(LD<EBCellFAB>)");
+  CH_TIMER("EBFineInterp::define_buffer", t1);
+  CH_TIMER("EBFineInterp::copyTo", t2);
+
+  CH_assert(m_isDefined);
+  CH_assert(a_fineData.nComp() > a_variables.end());
+  CH_assert(a_coarData.nComp() > a_variables.end());
+
+  CH_START(t1);
+  const DisjointBoxLayout& dblCoFi   = m_eblgCoFi.getDBL();
+  const EBISLayout&        ebislCoFi = m_eblgCoFi.getEBISL();
+  LevelData<EBCellFAB>     coarsenedFineData(dblCoFi, 1, IntVect::Zero, EBCellFactory(ebislCoFi));
+  CH_STOP(t1);
+
+  for (int ivar = a_variables.begin(); ivar <= a_variables.end(); ivar++) {
+    CH_START(t2);
[…]
