+++ Source/AmrMesh/CD_AmrMesh.H
@@ -1671,7 +1671,7 @@ public:
     @param[in] a_realm Realm name
     @param[in] a_phase Phase (gas or solid)
   */
-  Vector<RefCountedPtr<EBFineInterp>>&
+  Vector<RefCountedPtr<EBCoarseToFineInterp>>&
   getFineInterp(const std::string a_realm, const phase::which_phase a_phase) const;
 
   /*!

[…]
