+++ Source/AmrMesh/CD_EBCoarseToFineInterp.H
@@ -50,11 +50,11 @@ public:
     @param[in] a_ebisPtr  EB index space. 
   */
   EBCoarseToFineInterp(const EBLevelGrid&        a_eblgFine,
-               const EBLevelGrid&        a_eblgCoFi,
-               const EBLevelGrid&        a_eblgCoar,
-               const int&                a_refRat,
-               const int&                a_nComp,
-               const EBIndexSpace* const a_ebisPtr) noexcept;
+                       const EBLevelGrid&        a_eblgCoFi,
+                       const EBLevelGrid&        a_eblgCoar,
+                       const int&                a_refRat,
+                       const int&                a_nComp,
+                       const EBIndexSpace* const a_ebisPtr) noexcept;
 
   /*!
     @brief Disallowed copy constructor
@@ -102,38 +102,29 @@ public:
 
   /*!
     @brief Interpolate to the fine grid using the specified method. 
-    @param[out] a_fineData  Piecewise constant interpolation of the coarse data.
+    @param[out] a_fineData  Fine-grid data.
     @param[in]  a_coarData  Coarse data
     @param[in]  a_variables Variables to interpolate to the fine grid.
-    @note This simply calls the other methods. 
+    @param[in]  a_interpType Interpolation type
   */
   virtual void
-  interpolate(LevelData<EBCellFAB>&       a_fineData,
-              const LevelData<EBCellFAB>& a_coarData,
-              const Interval&             a_variables,
-              const EBCoarseToFineInterp::Type&   a_interpType) const noexcept;
+  interpolate(LevelData<EBCellFAB>&             a_fineData,
+              const LevelData<EBCellFAB>&       a_coarData,
+              const Interval&                   a_variables,
+              const EBCoarseToFineInterp::Type& a_interpType) const noexcept;
 
   /*!
-    @brief Do a piecewise constant interpolation to the new grids. This ignores the volume fraction of cut cells. 
-    @param[out] a_fineData  Piecewise constant interpolation of the coarse data.
+    @brief Interpolate to the fine grid using the specified method. 
+    @param[out] a_fineData  Fine-grid data
     @param[in]  a_coarData  Coarse data
     @param[in]  a_variables Variables to interpolate to the fine grid.
+    @param[in]  a_interpType Interpolation type
   */
   virtual void
-  regridArithmetic(LevelData<EBCellFAB>&       a_fineData,
-                   const LevelData<EBCellFAB>& a_coarData,
-                   const Interval&             a_variables) const noexcept;
-
-  /*!
-    @brief Do a piecewise constant interpolation to the new grids.
-    @param[out] a_fineData  Piecewise constant interpolation of the coarse data.
-    @param[in]  a_coarData  Coarse data
-    @param[in]  a_variables Variables to interpolate to the fine grid.
-  */
-  virtual void
-  regridConservativeNoSlopes(LevelData<EBCellFAB>&       a_fineData,
-                             const LevelData<EBCellFAB>& a_coarData,
-                             const Interval&             a_variables) const noexcept;
+  interpolate(LevelData<BaseIVFAB<Real>>&       a_fineData,
+              const LevelData<BaseIVFAB<Real>>& a_coarData,
+              const Interval&                   a_variables,
+              const EBCoarseToFineInterp::Type& a_interpType) const noexcept;
 
   /*!
     @brief Use min-mod slopes when interpolating to the new grids. 
@@ -143,28 +134,6 @@ public:
   */
   virtual void
   regridMinMod(LevelData<EBCellFAB>& a_fineData, const LevelData<EBCellFAB>& a_coarData, const Interval& a_variables);
-
-  /*!
-    @brief Regrid irregular data using conservative interpolation
[…]
