+++ Source/AmrMesh/CD_EBCoarseToFineInterp.cpp
@@ -27,11 +27,11 @@ EBCoarseToFineInterp::EBCoarseToFineInterp() noexcept
 }
 
 EBCoarseToFineInterp::EBCoarseToFineInterp(const EBLevelGrid&        a_eblgFine,
-                           const EBLevelGrid&        a_eblgCoFi,
-                           const EBLevelGrid&        a_eblgCoar,
-                           const int&                a_refRat,
-                           const int&                a_nComp,
-                           const EBIndexSpace* const a_ebisPtr) noexcept
+                                           const EBLevelGrid&        a_eblgCoFi,
+                                           const EBLevelGrid&        a_eblgCoar,
+                                           const int&                a_refRat,
+                                           const int&                a_nComp,
+                                           const EBIndexSpace* const a_ebisPtr) noexcept
 {
   CH_TIME("EBCoarseToFineInterp::EBCoarseToFineInterp(full)");
 
@@ -42,11 +42,11 @@ EBCoarseToFineInterp::~EBCoarseToFineInterp() noexcept {}
 
 void
 EBCoarseToFineInterp::define(const EBLevelGrid&        a_eblgFine,
-                     const EBLevelGrid&        a_eblgCoFi,		     
-                     const EBLevelGrid&        a_eblgCoar,
-                     const int&                a_refRat,
-                     const int&                a_nComp,
-                     const EBIndexSpace* const a_ebisPtr) noexcept
+                             const EBLevelGrid&        a_eblgCoFi,
+                             const EBLevelGrid&        a_eblgCoar,
+                             const int&                a_refRat,
+                             const int&                a_nComp,
+                             const EBIndexSpace* const a_ebisPtr) noexcept
 {
   CH_TIME("EBCoarseToFineInterp::define");
 
@@ -142,10 +142,10 @@ EBCoarseToFineInterp::defineWeights() noexcept
 }
 
 void
-EBCoarseToFineInterp::interpolate(LevelData<EBCellFAB>&       a_fineData,
-                          const LevelData<EBCellFAB>& a_coarData,
-                          const Interval&             a_variables,
-                          const EBCoarseToFineInterp::Type&   a_interpType) const noexcept
+EBCoarseToFineInterp::interpolate(LevelData<EBCellFAB>&             a_fineData,
+                                  const LevelData<EBCellFAB>&       a_coarData,
+                                  const Interval&                   a_variables,
+                                  const EBCoarseToFineInterp::Type& a_interpType) const noexcept
 {
   CH_TIMERS("EBCoarseToFineInterp::interpolate(LD<EBCellFAB>)");
   CH_TIMER("EBCoarseToFineInterp::define_buffer", t1);
@@ -169,17 +169,17 @@ EBCoarseToFineInterp::interpolate(LevelData<EBCellFAB>&       a_fineData,
     for (DataIterator dit(dblCoFi); dit.ok(); ++dit) {
       switch (a_interpType) {
       case EBCoarseToFineInterp::Type::Arithmetic: {
-        this->regridArithmetic(a_fineData, a_coarData, a_variables);
[…]
