### 2023-04-27 173f3ad
WIP. Changed signatures and about to transfer slopes to new class
