### 2023-04-27 0ebab21
Rename interpolation method
