+++ Physics/ItoKMC/CD_ItoKMCStepperImplem.H
@@ -5154,7 +5154,7 @@ ItoKMCStepper<I, C, R, F>::loadBalanceParticleRealm(Vector<Vector<int>>&        
       interpolator->interpolate(*speciesPPC[lvl],
                                 *speciesPPC[lvl - 1],
                                 Interval(0, 0),
-                                EBCoarseToFineInterp::Type::ConservativeNoSlopes);
+                                EBCoarseToFineInterp::Type::ConservativePWC);
 
       // There could be parts of the new grid that overlapped with the old grid (on level lvl) -- we don't want
       // to pollute the solution with interpolation there since we already have valid data.
