+++ Source/MeshODESolver/CD_MeshODESolverImplem.H
@@ -401,7 +401,7 @@ MeshODESolver<N>::regrid(const int a_lmin, const int a_oldFinestLevel, const int
   this->allocate();
 
   const EBCoarseToFineInterp::Type interpType = m_regridSlopes ? EBCoarseToFineInterp::Type::ConservativeMinMod
-                                                               : EBCoarseToFineInterp::Type::ConservativeNoSlopes;
+                                                               : EBCoarseToFineInterp::Type::ConservativePWC;
 
   m_amr->interpToNewGrids(m_phi, m_cache, m_phase, a_lmin, a_oldFinestLevel, a_newFinestLevel, interpType);
 
