+++ Source/ConvectionDiffusionReaction/CD_CdrSolver.cpp
@@ -1677,7 +1677,7 @@ CdrSolver::regrid(const int a_lmin, const int a_oldFinestLevel, const int a_newF
   this->allocate();
 
   const EBCoarseToFineInterp::Type interpType = m_regridSlopes ? EBCoarseToFineInterp::Type::ConservativeMinMod
-                                                               : EBCoarseToFineInterp::Type::ConservativeNoSlopes;
+                                                               : EBCoarseToFineInterp::Type::ConservativePWC;
 
   // Interpolate to the new grids.
   m_amr->interpToNewGrids(m_phi, m_cachePhi, m_phase, a_lmin, a_oldFinestLevel, a_newFinestLevel, interpType);
