+++ Source/SurfaceODESolver/CD_SurfaceODESolverImplem.H
@@ -597,8 +597,8 @@ SurfaceODESolver<N>::regrid(const int a_lmin, const int a_oldFinestLevel, const 
 
   this->allocate();
 
-  const EBCoarseToFineInterp::Type interpType = m_conservativeRegrid ? EBCoarseToFineInterp::Type::ConservativeNoSlopes
-                                                                     : EBCoarseToFineInterp::Type::Arithmetic;
+  const EBCoarseToFineInterp::Type interpType = m_conservativeRegrid ? EBCoarseToFineInterp::Type::ConservativePWC
+                                                                     : EBCoarseToFineInterp::Type::PWC;
 
   m_amr->interpToNewGrids(m_phi, m_cache, m_phase, a_lmin, a_oldFinestLevel, a_newFinestLevel, interpType);
 

[…]
