+++ Source/AmrMesh/CD_EBCoarseToFineInterp.H
@@ -13,16 +13,20 @@
 #define CD_EBCoarseToFineInterp_H
 
 // Chombo includes
-#include <RefCountedPtr.H>
 #include <EBLevelGrid.H>
-
-#include <EBPWLFineInterp.H>
+#include <LevelData.H>
+#include <EBCellFAB.H>
+#include <BaseIVFAB.H>
 
 // Our includes
 #include <CD_NamespaceHeader.H>
 
 /*!
   @brief Class for interpolating data to fine grids. Can use constant interpolation or include limiters.
+  @details This class should be used through the interpolate methods. The methods will switch between various supported
+  interpolation methods for the data types.
+  For EBCellFAB-type data we support PWC, ConservativePWC, ConservativeMinMod, ConservativeMonotonizedCentral, ConservativeSuperbee
+  For BASEIVFAB-type data we support PWC and ConservativePWC
 */
 class EBCoarseToFineInterp
 {
@@ -39,7 +43,8 @@ public:
     PWC,
     ConservativePWC,
     ConservativeMinMod,
-    ConservativeMonotizedCentral,
+    ConservativeMonotonizedCentral,
+    ConservativeSuperbee,
   };
 
   /*!
@@ -56,12 +61,10 @@ public:
     @param[in] a_nComp    Number of components. 
     @param[in] a_ebisPtr  EB index space. 
   */
-  EBCoarseToFineInterp(const EBLevelGrid&        a_eblgFine,
-                       const EBLevelGrid&        a_eblgCoFi,
-                       const EBLevelGrid&        a_eblgCoar,
-                       const int&                a_refRat,
-                       const int&                a_nComp,
-                       const EBIndexSpace* const a_ebisPtr) noexcept;
+  EBCoarseToFineInterp(const EBLevelGrid& a_eblgFine,
+                       const EBLevelGrid& a_eblgCoFi,
+                       const EBLevelGrid& a_eblgCoar,
+                       const int&         a_refRat) noexcept;
 
   /*!
     @brief Disallowed copy constructor
@@ -100,12 +103,10 @@ public:
     @param[in] a_ebisPtr  EB index space. 
   */
   virtual void
-  define(const EBLevelGrid&        a_eblgFine,
-         const EBLevelGrid&        a_eblgCoFi,
-         const EBLevelGrid&        a_eblgCoar,
-         const int&                a_refRat,
-         const int&                a_nComp,
-         const EBIndexSpace* const a_ebisPtr) noexcept;
+  define(const EBLevelGrid& a_eblgFine,
+         const EBLevelGrid& a_eblgCoFi,
+         const EBLevelGrid& a_eblgCoar,
+         const int&         a_refRat) noexcept;
 
   /*!
     @brief Interpolate to the fine grid using the specified method. 
@@ -135,6 +136,16 @@ public:
 
 protected:
   /*!
+    @brief Supported slope limiters for conservative interpolation with slopes.
+  */
+  enum class SlopeLimiter
+  {
+    MinMod,
+    MonotonizedCentral,
+    Superbee
+  };
+
+  /*!
     @brief Is defined or not
   */
   bool m_isDefined;
@@ -160,20 +171,16 @@ protected:
   EBLevelGrid m_eblgCoFi;
 
   /*!
-    @brief Irregular grid region.
+    @brief Weights in cut-cells when regridding conservatively
+    @details Defined over the fine layout.
   */
-  LayoutData<IntVectSet> m_irregRegions;
+  LevelData<BaseIVFAB<Real>> m_volumeWeights;
 
   /*!
     @brief Weights in cut-cells when regridding conservatively
     @details Defined over the fine layout. 
   */
[…]
