+++ Source/AmrMesh/CD_EBCoarseToFineInterp.cpp
@@ -9,6 +9,9 @@
   @author Robert Marskar
 */
 
+// Std includes
+#include <limits>
+
 // Chombo includes
 #include <CH_Timer.H>
 #include <EBCellFactory.H>
@@ -26,27 +29,23 @@ EBCoarseToFineInterp::EBCoarseToFineInterp() noexcept
   m_isDefined = false;
 }
 
-EBCoarseToFineInterp::EBCoarseToFineInterp(const EBLevelGrid&        a_eblgFine,
-                                           const EBLevelGrid&        a_eblgCoFi,
-                                           const EBLevelGrid&        a_eblgCoar,
-                                           const int&                a_refRat,
-                                           const int&                a_nComp,
-                                           const EBIndexSpace* const a_ebisPtr) noexcept
+EBCoarseToFineInterp::EBCoarseToFineInterp(const EBLevelGrid& a_eblgFine,
+                                           const EBLevelGrid& a_eblgCoFi,
+                                           const EBLevelGrid& a_eblgCoar,
+                                           const int&         a_refRat) noexcept
 {
   CH_TIME("EBCoarseToFineInterp::EBCoarseToFineInterp(full)");
 
-  this->define(a_eblgFine, a_eblgCoFi, a_eblgCoar, a_refRat, a_nComp, a_ebisPtr);
+  this->define(a_eblgFine, a_eblgCoFi, a_eblgCoar, a_refRat);
 }
 
 EBCoarseToFineInterp::~EBCoarseToFineInterp() noexcept {}
 
 void
-EBCoarseToFineInterp::define(const EBLevelGrid&        a_eblgFine,
-                             const EBLevelGrid&        a_eblgCoFi,
-                             const EBLevelGrid&        a_eblgCoar,
-                             const int&                a_refRat,
-                             const int&                a_nComp,
-                             const EBIndexSpace* const a_ebisPtr) noexcept
+EBCoarseToFineInterp::define(const EBLevelGrid& a_eblgFine,
+                             const EBLevelGrid& a_eblgCoFi,
+                             const EBLevelGrid& a_eblgCoar,
+                             const int&         a_refRat) noexcept
 {
   CH_TIMERS("EBCoarseToFineInterp::define");
   CH_TIMER("EBCoarseToFineInterp::define_ebpwlfineinterp", t0);
@@ -55,17 +54,6 @@ EBCoarseToFineInterp::define(const EBLevelGrid&        a_eblgFine,
   CH_assert(a_refRat > 1);
   CH_assert(a_nComp > 0);
   CH_assert(a_ebisPtr != nullptr);
-
-  CH_START(t0);
-  m_fineInterp = RefCountedPtr<EBPWLFineInterp>(new EBPWLFineInterp(a_eblgFine.getDBL(),
-                                                                    a_eblgCoar.getDBL(),
-                                                                    a_eblgFine.getEBISL(),
-                                                                    a_eblgCoar.getEBISL(),
-                                                                    a_eblgCoar.getDomain(),
-                                                                    a_refRat,
[…]
