### 2023-04-28 62274d2
Added a bunch of timers to EBGradient and put in correct function signatures. Might want to redo this class....
