+++ Source/AmrMesh/CD_EBGradient.cpp
@@ -34,7 +34,7 @@ EBGradient::EBGradient(const EBLevelGrid& a_eblg,
                        const Real         a_dx,
                        const int          a_refRat,
                        const int          a_order,
-                       const int          a_weighting)
+                       const int          a_weighting) noexcept
 {
   CH_TIME("EBGradient::EBGradient");
 
@@ -59,18 +59,9 @@ EBGradient::EBGradient(const EBLevelGrid& a_eblg,
     m_dxFine  = 1;
   }
 
-  ParmParse pp("EBGradient");
-  bool      profile     = false;
-  bool      forceNoEBCF = false;
-  pp.query("profile", profile);
-  pp.query("force_no_ebcf", forceNoEBCF);
-
-  Timer timer("EBGradient::EBGradient");
 
   // Define the level stencils. These are regular finite-difference stencils.
-  timer.startEvent("Define level stencils");
   this->defineLevelStencils();
-  timer.stopEvent("Define level stencils");
 
   // If there's finer level there might also be an EBCF crossing. Those can be tricky to deal with because
   // the stencil on this level (i.e., the coarse level) might reach underneath the finer level and obtain bogus data.
@@ -79,47 +70,37 @@ EBGradient::EBGradient(const EBLevelGrid& a_eblg,
   // where we can't find good stencils. If we do find such cells, we trigger m_hasEBCF which will define
   // a new set of grids where we run a least squares procedure. THOSE stencils are "dual-level" stencils
   // that ONLY reach into valid regions (i.e., not covered by a finer grid level).
-  if (m_hasFine && !forceNoEBCF) {
+  if (m_hasFine) {
 
     // Masks with transient lifetimes.
     LevelData<FArrayBox> coarMaskCF;
     LevelData<FArrayBox> coarMaskInvalid;
 
     // Define masks on the input grids.
-    timer.startEvent("EBCF masks");
     this->defineMasks(coarMaskCF, coarMaskInvalid);
-    timer.stopEvent("EBCF masks");
 
     // Define iterators for the input grid AND define the simplified buffer grids.
-    timer.startEvent("EBCF Iterators");
     this->defineIteratorsEBCF(coarMaskCF, coarMaskInvalid);
-    timer.stopEvent("EBCF Iterators");
 
     if (m_hasEBCF) {
       LevelData<FArrayBox> bufferCoarMaskCF;
       LevelData<FArrayBox> bufferCoarMaskInvalid;
 
-      timer.startEvent("EBCF buffers");
       this->defineBuffers(bufferCoarMaskCF, bufferCoarMaskInvalid, coarMaskCF, coarMaskInvalid);
-      timer.stopEvent("EBCF buffers");
-
-      timer.startEvent("EBCF stencils");
       this->defineStencilsEBCF(bufferCoarMaskInvalid);
-      timer.stopEvent("EBCF stencils");
     }
-  }
-
-  if (profile) {
-    timer.eventReport(pout(), false);
   }
 }
 
-EBGradient::~EBGradient() { CH_TIME("EBGradient::~EBGradient"); }
+EBGradient::~EBGradient() noexcept { CH_TIME("EBGradient::~EBGradient"); }
 
 void
-EBGradient::computeLevelGradient(LevelData<EBCellFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const
+EBGradient::computeLevelGradient(LevelData<EBCellFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const noexcept
 {
[…]
