+++ Source/AmrMesh/CD_EBGradient.H
@@ -73,15 +73,20 @@ public:
              const Real         a_dx,
              const int          a_refRat,
              const int          a_order,
-             const int          a_weighting = 0);
+             const int          a_weighting = 0) noexcept;
+
+  /*!
+    @brief Destructor (does nothing)
+  */
+  virtual ~EBGradient() noexcept;
 
   /*!
     @brief Compute gradient using data on the input level only. 
     @param[out] a_gradient Gradient of input scalar
     @param[in]  a_phi      Input scalar
   */
-  void
-  computeLevelGradient(LevelData<EBCellFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const;
+  virtual void
+  computeLevelGradient(LevelData<EBCellFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const noexcept;
 
   /*!
     @brief Compute the gradient component which is normal to the face, using centered differencing. 
@@ -91,8 +96,8 @@ public:
     @param[inout] a_gradient Gradient of input scalar
     @param[in]    a_phi      Input scalar
   */
-  void
-  computeNormalDerivative(LevelData<EBFluxFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const;
+  virtual void
+  computeNormalDerivative(LevelData<EBFluxFAB>& a_gradient, const LevelData<EBCellFAB>& a_phi) const noexcept;
 
   /*!
     @brief Compute gradient using two-level stencils (matching at EBCF). 
@@ -100,15 +105,10 @@ public:
     @param[in]  a_phi      Input scalar
     @param[in]  a_phiFine  Input scalar on finer level. 
   */
-  void
+  virtual void
   computeAMRGradient(LevelData<EBCellFAB>&       a_gradient,
                      const LevelData<EBCellFAB>& a_phi,
-                     const LevelData<EBCellFAB>& a_phiFine) const;
-
-  /*!
-    @brief Destructor (does nothing)
-  */
-  virtual ~EBGradient();
+                     const LevelData<EBCellFAB>& a_phiFine) const noexcept;
 
 protected:
   /*!
@@ -253,16 +253,16 @@ protected:
   /*!
     @brief Define level stencils. 
   */
-  void
-  defineLevelStencils();
+  virtual void
+  defineLevelStencils() noexcept;
 
   /*!
     @brief Make masks
     @param[out] a_coarMaskCF      Mask holding a value of 1 in coarse cells that live on the refinement boundary. 
     @param[out] a_coarMaskInvalid Mask holding a value of 1 in coarse cells that are covered by a finer grid level. 
   */
-  void
-  defineMasks(LevelData<FArrayBox>& a_coarMaskCF, LevelData<FArrayBox>& a_coarMaskInvalid);
+  virtual void
+  defineMasks(LevelData<FArrayBox>& a_coarMaskCF, LevelData<FArrayBox>& a_coarMaskInvalid) noexcept;
 
   /*!
     @brief Define buffers. 
@@ -273,26 +273,26 @@ protected:
     @param[in]  a_coarMaskCF            Mask for CF region on the coarse layout. 
     @param[in]  a_coarMaskInvalid       Mask for invalid regions on the coarse layout. 
   */
-  void
+  virtual void
   defineBuffers(LevelData<FArrayBox>&       a_bufferCoarMaskCF,
                 LevelData<FArrayBox>&       a_bufferCoarMaskInvalid,
                 const LevelData<FArrayBox>& a_coarMaskCF,
-                const LevelData<FArrayBox>& a_coarMaskInvalid);
+                const LevelData<FArrayBox>& a_coarMaskInvalid) noexcept;
 
   /*!
     @brief Define iterators and cells that need modified stencils
     @param[in]  a_coarMaskCF            Mask for CF region on the coarse layout. 
[…]
