### 2023-04-28 3086658
WIP before trying out new operator
