+++ Source/AmrMesh/CD_EBGradient.H
@@ -63,17 +63,21 @@ public:
     @brief Full constructor, assumes that there is a finer level. 
     @param[in] a_eblg         Grids
     @param[in] a_eblgFine     Fine grids
+    @param[in] a_eblgFiCo     Refinement of coarse grids
     @param[in] a_dx           Resolution
     @param[in] a_refRat       Refinement ratio
     @param[in] a_order        Interpolation order for modified gradients. 
     @param[in] a_weighting    Weighting factor least squares. 
+    @param[in] a_ghostVect    Number of ghost cells in the input data
   */
   EBGradient(const EBLevelGrid& a_eblg,
              const EBLevelGrid& a_eblgFine,
+             const EBLevelGrid& a_eblgFiCo,
              const Real         a_dx,
              const int          a_refRat,
              const int          a_order,
-             const int          a_weighting = 0) noexcept;
+             const int          a_weighting,
+             const IntVect      a_ghostVector) noexcept;
 
   /*!
     @brief Destructor (does nothing)
@@ -135,6 +139,11 @@ protected:
     @brief Fine grids
   */
   EBLevelGrid m_eblgFine;
+
+  /*!
+    @brief Refined coarse grids
+  */
+  EBLevelGrid m_eblgFiCo;
 
   /*!
     @brief True if we have a finer level (m_eblgFine is defined)
