### 2023-04-28 ec0a47d
Added define method. Will try to redo the operator now
