### 2023-04-28 95516f9
New EBGradient operator (#9)
