+++ Source/AmrMesh/CD_EBGradient.H
@@ -12,11 +12,9 @@
 #ifndef CD_EBGradient_H
 #define CD_EBGradient_H
 
-// Std includes
-#include <map>
-
 // Chombo includes
 #include <EBLevelGrid.H>
+#include <AggStencil.H>
 
 // Our includes
 #include <CD_Location.H>
@@ -64,6 +62,7 @@ public:
     @param[in] a_eblg         Grids
     @param[in] a_eblgFine     Fine grids
     @param[in] a_eblgFiCo     Refinement of coarse grids
+    @param[in] a_hasFine      Has fine level or not
     @param[in] a_dx           Resolution
     @param[in] a_refRat       Refinement ratio
     @param[in] a_order        Interpolation order for modified gradients. 
@@ -73,6 +72,7 @@ public:
   EBGradient(const EBLevelGrid& a_eblg,
              const EBLevelGrid& a_eblgFine,
              const EBLevelGrid& a_eblgFiCo,
+             const bool         a_hasFine,
              const Real         a_dx,
              const int          a_refRat,
              const int          a_order,
@@ -89,6 +89,7 @@ public:
     @param[in] a_eblg         Grids
     @param[in] a_eblgFine     Fine grids
     @param[in] a_eblgFiCo     Refinement of coarse grids
+    @param[in] a_hasFine      Has fine level or not
     @param[in] a_dx           Resolution
     @param[in] a_refRat       Refinement ratio
     @param[in] a_order        Interpolation order for modified gradients. 
@@ -99,6 +100,7 @@ public:
   define(const EBLevelGrid& a_eblg,
          const EBLevelGrid& a_eblgFine,
          const EBLevelGrid& a_eblgFiCo,
+         const bool         a_hasFine,
          const Real         a_dx,
          const int          a_refRat,
          const int          a_order,
@@ -207,24 +209,9 @@ protected:
   int m_refRat;
 
   /*!
-    @brief Coarse-grid buffers
+    @brief Number of ghost cells in incoming data holders
   */
-  DisjointBoxLayout m_bufferDblCoar;
-
-  /*!
-    @brief Fine-grid buffers
-  */
-  DisjointBoxLayout m_bufferDblFine;
-
-  /*!
-    @brief Coarse EBISLayout buffer
-  */
-  EBISLayout m_bufferEBISLCoar;
-
-  /*!
-    @brief Fine EBISLayout buffer
-  */
-  EBISLayout m_bufferEBISLFine;
+  IntVect m_ghostVector;
 
   /*!
     @brief Stencils for boundaries and cut-cells
@@ -234,15 +221,19 @@ protected:
 
   /*!
     @brief Part of EBCF gradient stencil which reaches into the coarse level. 
-    @note Lives on the buffer layout. 
   */
-  LayoutData<BaseIVFAB<VoFStencil>> m_bufferStencilsCoar;
+  LayoutData<BaseIVFAB<VoFStencil>> m_ebcfStencilsCoar;
 
   /*!
-    @brief Part of EBCF gradient stencil which reaches into the fine level. 
-    @note Lives on the buffer layout. 
+    @brief Part of EBCF gradient stencil which reaches into the coarse level. 
   */
-  LayoutData<BaseIVFAB<VoFStencil>> m_bufferStencilsFine;
+  LayoutData<BaseIVFAB<VoFStencil>> m_ebcfStencilsFine;
+
+  /*!
+    @brief AggStencils for boundaries and cut-cells. Provides faster application of m_levelStencils.
+    @note Lives on the coarse layout and only has one component. 
+  */
+  LayoutData<RefCountedPtr<AggStencil<EBCellFAB, EBCellFAB>>> m_aggLevelStencils[SpaceDim];
 
[…]
