+++ Source/AmrMesh/CD_EBGradient.cpp
@@ -20,6 +20,7 @@
 // Our includes
 #include <CD_Timer.H>
 #include <CD_EBGradient.H>
+#include <CD_ParallelOps.H>
 #include <CD_LeastSquares.H>
 #include <CD_VofUtils.H>
 #include <CD_LoadBalancing.H>
@@ -32,6 +33,7 @@ constexpr int EBGradient::m_nComp;
 EBGradient::EBGradient(const EBLevelGrid& a_eblg,
                        const EBLevelGrid& a_eblgFine,
                        const EBLevelGrid& a_eblgFiCo,
+                       const bool         a_hasFine,
                        const Real         a_dx,
                        const int          a_refRat,
                        const int          a_order,
@@ -40,13 +42,14 @@ EBGradient::EBGradient(const EBLevelGrid& a_eblg,
 {
   CH_TIME("EBGradient::EBGradient");
 
-  this->define(a_eblg, a_eblgFine, a_eblgFiCo, a_dx, a_refRat, a_order, a_weighting, a_ghostVector);
+  this->define(a_eblg, a_eblgFine, a_eblgFiCo, a_hasFine, a_dx, a_refRat, a_order, a_weighting, a_ghostVector);
 }
 
 void
 EBGradient::define(const EBLevelGrid& a_eblg,
                    const EBLevelGrid& a_eblgFine,
                    const EBLevelGrid& a_eblgFiCo,
+                   const bool         a_hasFine,
                    const Real         a_dx,
                    const int          a_refRat,
                    const int          a_order,
@@ -66,19 +69,23 @@ EBGradient::define(const EBLevelGrid& a_eblg,
   m_refRat       = a_refRat;
   m_order        = a_order;
   m_weighting    = a_weighting;
+  m_ghostVector  = a_ghostVector;
+  m_hasFine      = a_hasFine;
   m_hasEBCF      = false;
 
-  if (a_eblgFine.isDefined()) {
-    m_hasFine = true;
-    m_dxFine  = m_dx / m_refRat;
+  if (m_hasFine) {
+    m_dxFine = m_dx / m_refRat;
   }
   else {
-    m_hasFine = false;
-    m_dxFine  = 1;
+    m_dxFine = 1;
   }
 
   // Define the level stencils. These are regular finite-difference stencils.
   this->defineLevelStencils();
+
+  bool      makeEBCF = false;
+  ParmParse pp("EBGradient");
+  pp.query("ebcf", makeEBCF);
 
   // If there's finer level there might also be an EBCF crossing. Those can be tricky to deal with because
   // the stencil on this level (i.e., the coarse level) might reach underneath the finer level and obtain bogus data.
@@ -87,7 +94,7 @@ EBGradient::define(const EBLevelGrid& a_eblg,
   // where we can't find good stencils. If we do find such cells, we trigger m_hasEBCF which will define
   // a new set of grids where we run a least squares procedure. THOSE stencils are "dual-level" stencils
   // that ONLY reach into valid regions (i.e., not covered by a finer grid level).
-  if (m_hasFine) {
+  if (m_hasFine && makeEBCF) {
 
     // Masks with transient lifetimes.
     LevelData<FArrayBox> coarMaskCF;
@@ -100,13 +107,12 @@ EBGradient::define(const EBLevelGrid& a_eblg,
     this->defineIteratorsEBCF(coarMaskCF, coarMaskInvalid);
 
     if (m_hasEBCF) {
[…]
