+++ Source/Utilities/CD_DischargeIO.cpp
@@ -136,19 +136,191 @@ DischargeIO::writeEBHDF5Header(HDF5Handle&                a_handleH5,
 #ifdef CH_USE_HDF5
 void
 DischargeIO::writeEBHDF5Level(HDF5Handle&                 a_handleH5,
-                              const int                   a_level,
                               const LevelData<EBCellFAB>& a_outputData,
                               const ProblemDomain         a_domain,
                               const Real                  a_dx,
                               const Real                  a_dt,
                               const Real                  a_time,
+                              const int                   a_level,
                               const int                   a_refRatio,
                               const int                   a_numGhost) noexcept
 {
-  CH_TIME("DischargeIO::writeEBHDF5Level");
+  CH_TIMERS("DischargeIO::writeEBHDF5Level");
+  CH_TIMER("DischargeIO::writeEBHDF5Level::alloc", t1);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::copy_vars", t2);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::average_multicells", t3);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_regular", t4);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_covered", t5);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_iregular", t6);
 
   CH_assert(a_refRatio > 0);
   CH_assert(a_numGhost >= 0);
+
+  const int numInputVars      = a_outputData.nComp();
+  const int indexVolFrac      = numInputVars;
+  const int indexBoundaryArea = indexVolFrac + 1;
+  const int indexAreaFrac     = indexBoundaryArea + 1;
+  const int indexNormal       = indexAreaFrac + 2 * SpaceDim;
+  const int indexDist         = indexNormal + SpaceDim;
+  const int numCompTotal      = indexDist + 1;
+
+  const DisjointBoxLayout& dbl = a_outputData.disjointBoxLayout();
+
+  CH_START(t1);
+  LevelData<FArrayBox> levelData(dbl, numCompTotal, a_numGhost * IntVect::Unit);
+  CH_STOP(t1);
+
+  for (DataIterator dit(dbl); dit.ok(); ++dit) {
+    FArrayBox&       levelFAB      = levelData[dit()];
+    const EBCellFAB& outputData    = a_outputData[dit()];
+    const FArrayBox& outputDataReg = outputData.getFArrayBox();
+    const EBISBox&   ebisbox       = outputData.getEBISBox();
+    const Box        outputBox     = levelFAB.box() & a_domain;
+
+    // Copy the regular data. This will also do single-valued cut-cells. Later, we average the
+    // contents in multi-valued cells.
+    CH_START(t2);
+    levelFAB.setVal(0.);
+    levelFAB.copy(outputDataReg, 0, 0, numInputVars);
+    CH_STOP(t2);
+
+    // Run through the multi-valued cells and set the single-valued output-data to be the sum of the multi-valued data. I just
+    // don't know of a different way of doing this that would also play well with HDF5..
+    CH_START(t3);
+    const IntVectSet multiCells = ebisbox.getMultiCells(outputBox);
+    for (IVSIterator ivsIt(multiCells); ivsIt.ok(); ++ivsIt) {
+      const IntVect iv = ivsIt();
+
+      const Vector<VolIndex>& multiVofs = ebisbox.getVoFs(iv);
+      const int               numVoFs   = ebisbox.numVoFs(iv);
+
[…]
