+++ Source/Utilities/CD_DischargeIO.cpp
@@ -149,9 +149,9 @@ DischargeIO::writeEBHDF5Level(HDF5Handle&                 a_handleH5,
   CH_TIMER("DischargeIO::writeEBHDF5Level::alloc", t1);
   CH_TIMER("DischargeIO::writeEBHDF5Level::copy_vars", t2);
   CH_TIMER("DischargeIO::writeEBHDF5Level::average_multicells", t3);
-  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_regular", t4);
-  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_covered", t5);
-  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_iregular", t6);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_default_data", t4);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_eb_moments", t5);
+  CH_TIMER("DischargeIO::writeEBHDF5Level::set_ghosts", t6);
 
   CH_assert(a_refRatio > 0);
   CH_assert(a_numGhost >= 0);
@@ -210,50 +210,42 @@ DischargeIO::writeEBHDF5Level(HDF5Handle&                 a_handleH5,
     }
     CH_STOP(t3);
 
+    CH_START(t4);
     // Set default volume fraction, area fraction, normal, and distance of EB from corner.
-    const bool isAllCovered = ebisbox.isAllCovered();
-    const bool isAllRegular = ebisbox.isAllRegular();
-
-    if (isAllRegular) {
-      CH_START(t4);
-      levelFAB.setVal(1.0, indexVolFrac);
-      levelFAB.setVal(0.0, indexBoundaryArea);
-      levelFAB.setVal(0.0, indexDist);
-
-      for (int i = 0; i < 2 * SpaceDim; i++) {
-        levelFAB.setVal(1.0, indexAreaFrac + i);
-      }
-
-      for (int i = 0; i < SpaceDim; i++) {
-        levelFAB.setVal(0.0, indexNormal + i);
-      }
-      CH_STOP(t4);
+    levelFAB.setVal(1.0, indexVolFrac);
+    levelFAB.setVal(0.0, indexBoundaryArea);
+    levelFAB.setVal(0.0, indexDist);
+    for (int i = 0; i < 2 * SpaceDim; i++) {
+      levelFAB.setVal(1.0, indexAreaFrac + i);
     }
-    else if (isAllCovered) {
-      CH_START(t5);
-      levelFAB.setVal(0.0, indexVolFrac);
-      levelFAB.setVal(0.0, indexBoundaryArea);
-      levelFAB.setVal(0.0, indexDist);
 
-      for (int i = 0; i < 2 * SpaceDim; i++) {
-        levelFAB.setVal(0.0, indexAreaFrac + i);
-      }
-
-      for (int i = 0; i < SpaceDim; i++) {
-        levelFAB.setVal(0.0, indexNormal + i);
-      }
-      CH_STOP(t5);
+    for (int i = 0; i < SpaceDim; i++) {
+      levelFAB.setVal(0.0, indexNormal + i);
     }
-    else {
-      CH_START(t6);
-      const IntVectSet irregIVS = ebisbox.getIrregIVS(outputBox);
+    CH_STOP(t4);
 
-      for (IVSIterator ivsIt(irregIVS); ivsIt.ok(); ++ivsIt) {
-        const IntVect iv = ivsIt();
+    // Set EB moment data for each cell.
+    CH_START(t5);
+    for (BoxIterator bit(outputBox); bit.ok(); ++bit) {
+      const IntVect& iv = bit();
+
+      const bool isCovered   = ebisbox.isCovered(iv);
[…]
