+++ Source/Utilities/CD_DischargeIO.cpp
@@ -386,10 +386,8 @@ DischargeIO::writeEBHDF5(const std::string&                   a_filename,
   // write header data
   DischargeIO::writeEBHDF5Header(handle, a_numLevels, a_probLo, a_variableNames);
 
-  // Set things up for each level
+  // Write levfle-by-level
   for (int lvl = 0; lvl < a_numLevels; lvl++) {
-#if 1
-    // Write this level.
     const int refRat = (lvl < a_numLevels) ? a_refinementRatios[lvl] : 1;
     DischargeIO::writeEBHDF5Level(handle,
                                   *a_data[lvl],
@@ -400,182 +398,6 @@ DischargeIO::writeEBHDF5(const std::string&                   a_filename,
                                   lvl,
                                   refRat,
                                   a_numGhost);
-#else
-    const DisjointBoxLayout& dbl = a_grids[lvl];
-
-    // This is the domain and grid resolution on this level.
-    const Real          dx     = a_dx[lvl];
-    const ProblemDomain domain = a_domains[lvl];
-
-    // Allocate the Chombo FArrayBox data.
-    LevelData<FArrayBox> levelData(dbl, numCompTotal, a_numGhost * IntVect::Unit);
-
-    // Go through all the grids on this level
-    for (DataIterator dit(dbl); dit.ok(); ++dit) {
-      // Handle to single-valued data.
-      FArrayBox& currentFab = levelData[dit()];
-
-      // Grid information and user input variables.
-      const EBCellFAB& data    = (*a_data[lvl])[dit()];
-      const EBISBox&   ebisbox = data.getEBISBox();
-      const Box        box     = currentFab.box() & domain;
-
-      // Copy the regular data. This will also do single-valued cut-cells.
-      constexpr int srcStartComp = 0;
-      constexpr int dstStartComp = 0;
-
-      currentFab.setVal(0.);
-      currentFab.copy(data.getSingleValuedFAB(), srcStartComp, dstStartComp, numInputVars);
-
-      // Run through the multi-valued cells and set the single-valued output-data to be the sum of the multi-valued data. I just
-      // don't know of a different way of doing this that would also play well with HDF5..
-      const IntVectSet multiCells = ebisbox.getMultiCells(currentFab.box());
-      for (IVSIterator ivsIt(multiCells); ivsIt.ok(); ++ivsIt) {
-        const IntVect iv = ivsIt();
-
-        const Vector<VolIndex>& multiVofs = ebisbox.getVoFs(iv);
-        const int               numVoFs   = ebisbox.numVoFs(iv);
-
-        const Real invNumVoFs = 1. / (Real(numVoFs));
-
-        // Set to average
-        for (int comp = 0; comp < numInputVars; ++comp) {
-          currentFab(iv, comp) = 0.0;
-
-          for (int ivof = 0; ivof < numVoFs; ivof++) {
-            currentFab(iv, comp) += data(multiVofs[ivof], comp);
-          }
-
-          currentFab(iv, comp) *= invNumVoFs;
-        }
-      }
-
-      // Set default volume fraction, area fraction, normal, and distance of EB from corner.
-      currentFab.setVal(1.0, indexVolFrac);
-      currentFab.setVal(0.0, indexBoundaryArea);
-      currentFab.setVal(0.0, indexDist);
-
-      for (int i = 0; i < 2 * SpaceDim; i++) {
-        currentFab.setVal(1.0, indexAreaFrac + i);
-      }
-
-      for (int i = 0; i < SpaceDim; i++) {
-        currentFab.setVal(0.0, indexNormal + i);
-      }
-
-      // Run through the grid box and set the grid information.
-      for (BoxIterator bit(box); bit.ok(); ++bit) {
-        const IntVect& iv = bit();
-
[…]
