+++ Source/Driver/CD_Driver.cpp
@@ -1026,7 +1026,7 @@ Driver::parseOptions()
   pp.get("max_steps", m_maxSteps);
   pp.get("start_time", m_startTime);
   pp.get("stop_time", m_stopTime);
-  pp.get("max_plot_depth", m_maxPlotDepth);
+  pp.get("max_plot_depth", m_maxPlotLevel);
   pp.get("max_chk_depth", m_maxCheckpointDepth);
   pp.get("do_init_load_balance", m_doInitLoadBalancing);
 
@@ -2130,125 +2130,107 @@ Driver::writeCrashFile()
 void
 Driver::writePlotFile(const std::string a_filename)
 {
-  CH_TIME("Driver::writePlotFile(string)");
+  CH_TIMERS("Driver::writePlotFile(string)");
+  CH_TIMER("Driver::writePlotFile::allocate", t1);
+  CH_TIMER("Driver::writePlotFile::assemble", t2);
+  CH_TIMER("Driver::writePlotFile::interp_exchange", t3);
+  CH_TIMER("Driver::writePlotFile::copy_internal", t4);
+  CH_TIMER("Driver::writePlotFile::h5_write", t5);
+
   if (m_verbosity > 3) {
     pout() << "Driver::writePlotFile(string)" << endl;
   }
 
-  // TLDR: This is the main routine for writing plot files. It consists of a bit of code with two essential components:
-  //
-  //       1. Copy data over from various solvers/celltaggers etc. and put them in a single EBAMRCellData with the
-  //          necessary amount of variables.
-  //
-  //       2. Write the big EBAMRCellData to file.
-  //
-  // For performance tracking, this routine can be timed.
-
-  Timer timer("Driver::writePlotFile(string)");
-
-  // Output file
-  EBAMRCellData output;
-  EBAMRCellData scratch;
-
-  // Names for output variables
-  Vector<std::string> plotVariableNames(0);
-
   // Get total number of components for output
-  int ncomp = m_timeStepper->getNumberOfPlotVariables();
+  int numOutputComp = m_timeStepper->getNumberOfPlotVariables();
   if (!m_cellTagger.isNull()) {
-    ncomp += m_cellTagger->getNumberOfPlotVariables();
+    numOutputComp += m_cellTagger->getNumberOfPlotVariables();
   }
-  ncomp += this->getNumberOfPlotVariables();
+  numOutputComp += this->getNumberOfPlotVariables();
 
   // If there's nothing to write, return.
-  if (ncomp > 0) {
+  if (numOutputComp > 0) {
+    Timer timer("Driver::writePlotFile(string)");
+
+    Vector<std::string> plotVariableNames(0);
 
     // Allocate storage
-    m_amr->allocate(output, m_realm, phase::gas, ncomp);
-    m_amr->allocate(scratch, m_realm, phase::gas, 1);
+    CH_START(t1);
+    EBAMRCellData output;
+    m_amr->allocate(output, m_realm, phase::gas, numOutputComp);
     DataOps::setValue(output, 0.0);
-    DataOps::setValue(scratch, 0.0);
-
-    // Assemble data
-    int icomp = 0; // Used as reference for output components
+    CH_STOP(t1);
 
     timer.startEvent("Assemble data");
     if (m_verbosity >= 3) {
       pout() << "Driver::writePlotFile - assembling data..." << endl;
     }
 
-    // Time stepper writes its data
-    m_timeStepper->writePlotData(output, plotVariableNames, icomp);
+    CH_START(t2);
 
-    // Cell tagger writes data
+    int icomp = 0;
+    m_timeStepper->writePlotData(output, plotVariableNames, icomp);
     if (!m_cellTagger.isNull()) {
       m_cellTagger->writePlotData(output, plotVariableNames, icomp);
     }
-
[…]
