### 2023-05-02 c067167
updated rod/needle geometry to include the wider part of the rod
