### 2023-05-02 fdd36a4
moved m_prevDt=dt to advance-method
