### 2023-05-02 0aa9d67
I/O memory reduction (#10)
