+++ Source/MeshODESolver/CD_MeshODESolverImplem.H
@@ -448,54 +448,6 @@ MeshODESolver<N>::setName(const std::string& a_name) noexcept
 }
 
 template <size_t N>
-Vector<std::string>
-MeshODESolver<N>::getPlotVariableNames() const noexcept
-{
-  CH_TIME("MeshODESolver::getPlotVariableNames()");
-  if (m_verbosity > 5) {
-    pout() << m_name + "::getPlotVariableNames()" << endl;
-  }
-
-  Vector<std::string> plotVarNames(0);
-
-  if (m_plotPhi) {
-    for (size_t i = 0; i < N; i++) {
-      plotVarNames.push_back(m_name + " phi-" + std::to_string(i));
-    }
-  }
-
-  if (m_plotRHS) {
-    for (size_t i = 0; i < N; i++) {
-      plotVarNames.push_back(m_name + " source-" + std::to_string(i));
-    }
-  }
-
-  return plotVarNames;
-}
-
-template <size_t N>
-int
-MeshODESolver<N>::getNumberOfPlotVariables() const noexcept
-{
-  CH_TIME("MeshODESolver::getNumberOfPlotVariables()");
-  if (m_verbosity > 5) {
-    pout() << m_name + "::getNumberOfPlotVariables()" << endl;
-  }
-
-  int numPlotVars = 0;
-
-  if (m_plotPhi) {
-    numPlotVars += int(N);
-  }
-
-  if (m_plotRHS) {
-    numPlotVars += int(N);
-  }
-
-  return numPlotVars;
-}
-
-template <size_t N>
 void
 MeshODESolver<N>::registerOperators() const noexcept
 {
@@ -580,7 +532,9 @@ MeshODESolver<N>::writePlotFile() const noexcept
   DataOps::setValue(output, 0.0);
 
   int icomp = 0;
-  this->writePlotData(output, icomp);
+  for (int lvl = 0; lvl <= m_amr->getFinestLevel(); lvl++) {
+    this->writePlotData(*output[lvl], icomp, lvl);
+  }
 
   char filename[100];
   sprintf(filename, "%s.step%07d.%dd.hdf5", m_name.c_str(), m_timeStep, SpaceDim);
@@ -609,21 +563,147 @@ MeshODESolver<N>::writePlotFile() const noexcept
 }
 
 template <size_t N>
-void
-MeshODESolver<N>::writePlotData(EBAMRCellData& a_output, int& a_icomp) const noexcept
+int
+MeshODESolver<N>::getNumberOfPlotVariables() const noexcept
 {
-  CH_TIME("MeshODESolver::writePlotData(EBAMRCellData, int)");
+  CH_TIME("MeshODESolver::getNumberOfPlotVariables()");
   if (m_verbosity > 5) {
-    pout() << m_name + "::writePlotData(EBAMRCellData, int)" << endl;
+    pout() << m_name + "::getNumberOfPlotVariables()" << endl;
   }
 
+  int numPlotVars = 0;
+
   if (m_plotPhi) {
-    this->writeData(a_output, a_icomp, m_phi, false);
+    numPlotVars += int(N);
   }
 
   if (m_plotRHS) {
-    this->writeData(a_output, a_icomp, m_rhs, false);
+    numPlotVars += int(N);
   }
+
+  return numPlotVars;
+}
+
+template <size_t N>
+Vector<std::string>
+MeshODESolver<N>::getPlotVariableNames() const noexcept
+{
+  CH_TIME("MeshODESolver::getPlotVariableNames()");
+  if (m_verbosity > 5) {
+    pout() << m_name + "::getPlotVariableNames()" << endl;
+  }
+
+  Vector<std::string> plotVarNames(0);
+
[…]

[…]
