+++ Source/AmrMesh/CD_EBGhostCellInterpolator.cpp
@@ -1,0 +1,468 @@
+/* chombo-discharge
+ * Copyright © 2023 SINTEF Energy Research.
+ * Please refer to Copyright.txt and LICENSE in the chombo-discharge root directory.
+ */
+
+/*!
+  @file   CD_EBGhostCellInterpolator.cpp
+  @brief  Implementation of CD_EBGhostCellInterpolator.H
+  @author Robert Marskar
+*/
+
+// Chombo includes
+#include <CH_Timer.H>
+#include <IntVectSet.H>
+#include <EBCellFactory.H>
+#include <EBAlias.H>
+#include <NeighborIterator.H>
+
+// Our includes
+#include <CD_EBGhostCellInterpolator.H>
+#include <CD_BoxLoops.H>
+#include <CD_NamespaceHeader.H>
+
+EBGhostCellInterpolator::EBGhostCellInterpolator() noexcept
+{
+  CH_TIME("EBGhostCellInterpolator::EBGhostCellInterpolator(weak)");
+
+  m_isDefined = false;
+}
+
+EBGhostCellInterpolator::EBGhostCellInterpolator(const EBLevelGrid& a_eblgFine,
+                                                 const EBLevelGrid& a_eblgCoFi,
+                                                 const EBLevelGrid& a_eblgCoar,
+                                                 const IntVect&     a_ghostVector,
+                                                 const int          a_refRat,
+                                                 const int          a_ghostCF) noexcept
+{
+  CH_TIME("EBGhostCellInterpolator::EBGhostCellInterpolator(strong)");
+
+  this->define(a_eblgFine, a_eblgCoFi, a_eblgCoar, a_ghostVector, a_refRat, a_ghostCF);
+}
+
+EBGhostCellInterpolator::~EBGhostCellInterpolator() noexcept
+{
+  CH_TIME("EBGhostCellInterpolator::~EBGhostCellInterpolator");
+}
+
+void
+EBGhostCellInterpolator::define(const EBLevelGrid& a_eblgFine,
+                                const EBLevelGrid& a_eblgCoFi,
+                                const EBLevelGrid& a_eblgCoar,
+                                const IntVect&     a_ghostVector,
+                                const int          a_refRat,
+                                const int          a_ghostCF) noexcept
+{
+  CH_TIME("EBGhostCellInterpolator::define()");
+
+  CH_assert(a_ghostCF >= 0);
+  CH_assert(a_refRat >= 2);
+
+  m_eblgFine    = a_eblgFine;
+  m_eblgCoFi    = a_eblgCoFi;
+  m_eblgCoar    = a_eblgCoar;
+  m_ghostVector = a_ghostVector;
+  m_refRat      = a_refRat;
+  m_ghostCF     = a_ghostCF;
+
+  this->defineGhostRegions();
+
+  m_isDefined = true;
+}
+
+void
+EBGhostCellInterpolator::defineGhostRegions() noexcept
+{
+  CH_TIME("EBGhostCellInterpolator::defineGhostRegions");
+
+  const DisjointBoxLayout& dblFine = m_eblgFine.getDBL();
+  const DisjointBoxLayout& dblCoar = m_eblgCoFi.getDBL();
+
+  const EBISLayout& ebislFine = m_eblgFine.getEBISL();
+  const EBISLayout& ebislCoar = m_eblgCoFi.getEBISL();
+
+  const ProblemDomain& domainFine = m_eblgFine.getDomain();
+  const ProblemDomain& domainCoar = m_eblgCoFi.getDomain();
+
+  m_regularGhostRegions.define(dblFine);
+  m_fineIrregCells.define(dblFine);
+  m_coarIrregCells.define(dblCoar);
[…]
