+++ Source/AmrMesh/CD_PhaseRealm.cpp
@@ -632,7 +632,7 @@ PhaseRealm::defineFillPatch(const int a_lmin)
 
   const bool doThisOperator = this->queryOperator(s_eb_fill_patch);
 
-  m_pwlFillPatch.resize(1 + m_finestLevel);
+  m_ghostCellInterpolator.resize(1 + m_finestLevel);
 
   if (doThisOperator) {
 
@@ -645,17 +645,13 @@ PhaseRealm::defineFillPatch(const int a_lmin)
 
       // Filling ghost cells on level l from coarse data on level l-1 is stored on level l
       if (hasCoar) {
-        m_pwlFillPatch[lvl] = RefCountedPtr<AggEBPWLFillPatch>(new AggEBPWLFillPatch(m_grids[lvl],
-                                                                                     m_grids[lvl - 1],
-                                                                                     m_ebisl[lvl],
-                                                                                     m_ebisl[lvl - 1],
-                                                                                     m_domains[lvl - 1],
-                                                                                     m_refinementRatios[lvl - 1],
-                                                                                     comps,
-                                                                                     radius,
-                                                                                     ghost,
-                                                                                     !m_hasEbCf,
-                                                                                     &(*m_ebis)));
+        m_ghostCellInterpolator[lvl] = RefCountedPtr<EBGhostCellInterpolator>(
+          new EBGhostCellInterpolator(*m_eblg[lvl],
+                                      *m_eblgCoFi[lvl - 1],
+                                      *m_eblg[lvl - 1],
+                                      ghost,
+                                      m_refinementRatios[lvl - 1],
+                                      radius));
       }
     }
   }
@@ -1139,14 +1135,14 @@ PhaseRealm::getMultigridInterpolator() const
   return m_multigridInterpolator;
 }
 
-Vector<RefCountedPtr<AggEBPWLFillPatch>>&
-PhaseRealm::getFillPatch() const
+Vector<RefCountedPtr<EBGhostCellInterpolator>>&
+PhaseRealm::getGhostCellInterpolator() const
 {
   if (!this->queryOperator(s_eb_fill_patch)) {
     MayDay::Error("PhaseRealm::getFillPatch - operator not registered!");
   }
 
-  return m_pwlFillPatch;
+  return m_ghostCellInterpolator;
 }
 
 Vector<RefCountedPtr<EBCoarseToFineInterp>>&
