### 2023-05-09 4feff36
More efficient table resampling
