### 2023-05-09 be4f098
Explicit stencils to coarsening utility (#13)
