+++ Source/Electrostatics/CD_FieldSolverMultigrid.cpp
@@ -226,32 +226,16 @@ FieldSolverMultigrid::parseJumpBC()
   }
 }
 
-void
-FieldSolverMultigrid::allocate()
-{
-  CH_TIME("FieldSolverMultigrid::allocate()");
-  if (m_verbosity > 5) {
-    pout() << "FieldSolverMultigrid::allocate()" << endl;
-  }
-
-  FieldSolver::allocate();
-
-  m_amr->allocate(m_zero, m_realm, m_nComp);
-  m_amr->allocate(m_kappaRhoByEps0, m_realm, m_nComp);
-  m_amr->allocate(m_sigmaByEps0, m_realm, phase::gas, m_nComp);
-
-  DataOps::setValue(m_zero, 0.0);
-  DataOps::setValue(m_kappaRhoByEps0, 0.0);
-  DataOps::setValue(m_sigmaByEps0, 0.0);
-}
-
 bool
 FieldSolverMultigrid::solve(MFAMRCellData&       a_phi,
                             const MFAMRCellData& a_rho,
                             const EBAMRIVData&   a_sigma,
                             const bool           a_zeroPhi)
 {
-  CH_TIME("FieldSolverMultigrid::solve(MFAMRCellData, MFAMRCellData, EBAMRIVData, bool)");
+  CH_TIMERS("FieldSolverMultigrid::solve");
+  CH_TIMER("FieldSolverMultigrid::solve::alloc_temps", t1);
+  CH_TIMER("FieldSolverMultigrid::solve::set_temps", t2);
+  CH_TIMER("FieldSolverMultigrid::solve::smooth_phi", t3);
   if (m_verbosity > 5) {
     pout() << "FieldSolverMultigrid::solve(MFAMRCellData, MFAMRCellData, EBAMRIVData, bool)" << endl;
   }
@@ -269,42 +253,60 @@ FieldSolverMultigrid::solve(MFAMRCellData&       a_phi,
 
   bool converged = false;
 
+  // Set up multigrid solver if it is not already done.
   if (!m_isSolverSetup) {
-    this->setupSolver(); // This does everything, allocates coefficients, gets bc stuff and so on
+    this->setupSolver();
   }
 
+  // Define temporaries; the incoming data might need to be scaled but we don't want to
+  // alter it directly.
+  CH_START(t1);
+  MFAMRCellData zero;
+  MFAMRCellData kappaRhoByEps0;
+  EBAMRIVData   sigmaByEps0;
+
+  m_amr->allocate(zero, m_realm, m_nComp);
+  m_amr->allocate(kappaRhoByEps0, m_realm, m_nComp);
+  m_amr->allocate(sigmaByEps0, m_realm, phase::gas, m_nComp);
+  CH_STOP(t1);
+
+  // Scale data as appropriate.
+  CH_START(t2);
+  DataOps::setValue(zero, 0.0);
+
   // Do the scaled space charge density.
-  DataOps::copy(m_kappaRhoByEps0, a_rho);
-  DataOps::scale(m_kappaRhoByEps0, 1. / (Units::eps0));
+  DataOps::copy(kappaRhoByEps0, a_rho);
+  DataOps::scale(kappaRhoByEps0, 1. / (Units::eps0));
 
   // Special flag for when a_rho is on the centroid but was not scaled by kappa on input. The multigrid operator solves
   // kappa*L(phi) = kappa*rho so the right-hand side must be kappa-weighted.
   if (m_kappaSource) {
-    DataOps::kappaScale(m_kappaRhoByEps0);
+    DataOps::kappaScale(kappaRhoByEps0);
   }
 
   m_amr->conservativeAverage(a_phi, m_realm);
[…]
