+++ Source/AmrMesh/CD_EBCoarAve.cpp
@@ -1,5 +1,5 @@
 /* chombo-discharge
- * Copyright © 2021 SINTEF Energy Research.
+ * Copyright © 2023 SINTEF Energy Research.
  * Please refer to Copyright.txt and LICENSE in the chombo-discharge root directory.
  */
 
@@ -84,7 +84,235 @@ EBCoarAve::define(const EBLevelGrid& a_eblgFine,
     m_irregSetsCoFi[dit()] = m_eblgCoFi.getEBISL()[dit()].getIrregIVS(m_eblgCoFi.getDBL().get(dit()));
   }
 
+  this->defineCellStencils();
+  this->defineFaceStencils();
+  this->defineEBStencils();
+
   m_isDefined = true;
+}
+
+void
+EBCoarAve::defineCellStencils() noexcept
+{
+  CH_TIME("EBCoarAve::defineCellStencils");
+
+  const DisjointBoxLayout& dblCoar = m_eblgCoFi.getDBL();
+  const DisjointBoxLayout& dblFine = m_eblgFine.getDBL();
+
+  const EBISLayout& ebislCoar = m_eblgCoFi.getEBISL();
+  const EBISLayout& ebislFine = m_eblgFine.getEBISL();
+
+  const Real dxCoar   = 1.0;
+  const Real dxFine   = dxCoar / m_refRat;
+  const Real dxFactor = std::pow(dxFine / dxCoar, SpaceDim);
+
+  m_irregCellsCoFi.define(dblCoar);
+  m_cellConservativeStencils.define(dblCoar);
+  m_cellArithmeticStencils.define(dblCoar);
+  m_cellHarmonicStencils.define(dblCoar);
+
+  for (DataIterator dit(dblCoar); dit.ok(); ++dit) {
+    const EBISBox& ebisBoxCoar = ebislCoar[dit()];
+    const EBISBox& ebisBoxFine = ebislFine[dit()];
+    const EBGraph& ebGraphCoar = ebisBoxCoar.getEBGraph();
+
+    VoFIterator&           irregCells           = m_irregCellsCoFi[dit()];
+    BaseIVFAB<VoFStencil>& conservativeStencils = m_cellConservativeStencils[dit()];
+    BaseIVFAB<VoFStencil>& arithmeticStencils   = m_cellArithmeticStencils[dit()];
+    BaseIVFAB<VoFStencil>& harmonicStencils     = m_cellHarmonicStencils[dit()];
+
+    irregCells.define(m_irregSetsCoFi[dit()], ebGraphCoar);
+    conservativeStencils.define(m_irregSetsCoFi[dit()], ebGraphCoar, 1);
+    arithmeticStencils.define(m_irregSetsCoFi[dit()], ebGraphCoar, 1);
+    harmonicStencils.define(m_irregSetsCoFi[dit()], ebGraphCoar, 1);
+
+    auto buildStencils = [&](const VolIndex& coarVoF) -> void {
+      const Real             kappaC      = ebisBoxCoar.volFrac(coarVoF);
+      const Vector<VolIndex> fineVoFs    = ebislCoar.refine(coarVoF, m_refRat, dit());
+      const int              numFineVoFs = fineVoFs.size();
+
+      VoFStencil& arithSten = arithmeticStencils(coarVoF, 0);
+      VoFStencil& harmSten  = harmonicStencils(coarVoF, 0);
+      VoFStencil& consSten  = conservativeStencils(coarVoF, 0);
+
+      arithSten.clear();
+      harmSten.clear();
+      consSten.clear();
+
+      if (numFineVoFs > 0) {
+
+        for (int ifine = 0; ifine < numFineVoFs; ifine++) {
[…]
