+++ Source/Elliptic/CD_MFHelmholtzOp.cpp
@@ -489,7 +489,6 @@ void
 MFHelmholtzOp::preCond(LevelData<MFCellFAB>& a_corr, const LevelData<MFCellFAB>& a_residual)
 {
   CH_TIME("MFHelmholtzOp::preCond(LD<MFCellFAB>, LD<MFCellFAB>)");
-
 #if 1
   this->relax(a_corr, a_residual, 40);
 #else
@@ -1005,7 +1004,7 @@ MFHelmholtzOp::AMRResidual(LevelData<MFCellFAB>&             a_residual,
                            bool                              a_homogeneousPhysBC,
                            AMRLevelOp<LevelData<MFCellFAB>>* a_finerOp)
 {
-  CH_TIME("MFHelmholtzOp::AMRResidual(5x LD<MFCellFAB>, bool, AMRLevelOp<LevelData<MFCellFAB> >)");
+  CH_TIME("MFHelmholtzOp::AMRResidual");
 
   // Make residual = a_rhs - L(phi)
   this->AMROperator(a_residual, a_phiFine, a_phi, a_phiCoar, a_homogeneousPhysBC, a_finerOp);
@@ -1021,7 +1020,7 @@ MFHelmholtzOp::AMRResidualNF(LevelData<MFCellFAB>&       a_residual,
                              const LevelData<MFCellFAB>& a_rhs,
                              bool                        a_homogeneousPhysBC)
 {
-  CH_TIME("MFHelmholtzOp::AMRResidualNF(LD<MFCellFAB>, LD<MFCellFAB>, LD<MFCellFAB>, LD<MFCellFAB>, bool)");
+  CH_TIME("MFHelmholtzOp::AMRResidualNF");
 
   // Make residual = a_rhs - L(phi)
   this->AMROperatorNF(a_residual, a_phi, a_phiCoar, a_homogeneousPhysBC);
@@ -1038,8 +1037,7 @@ MFHelmholtzOp::AMRResidualNC(LevelData<MFCellFAB>&             a_residual,
                              bool                              a_homogeneousPhysBC,
                              AMRLevelOp<LevelData<MFCellFAB>>* a_finerOp)
 {
-  CH_TIME(
-    "MFHelmholtzOp::AMRResidualNC(LD<MFCellFAB>, LD<MFCellFAB>, LD<MFCellFAB>, LD<MFCellFAB>, bool, AMRLevelOp<LevelData<MFCellFAB>)");
+  CH_TIME("MFHelmholtzOp::AMRResidualNC");
 
   this->AMROperatorNC(a_residual, a_phiFine, a_phi, a_homogeneousPhysBC, a_finerOp);
 
@@ -1053,14 +1051,11 @@ MFHelmholtzOp::AMROperatorNF(LevelData<MFCellFAB>&       a_Lphi,
                              const LevelData<MFCellFAB>& a_phiCoar,
                              bool                        a_homogeneousPhysBC)
 {
-  CH_TIME("MFHelmholtzOp::AMROperatorNF(LD<MFCellFAB>, LD<MFCellFAB>, LD<MFCellFAB>, bool)");
+  CH_TIME("MFHelmholtzOp::AMROperatorNF");
 
   constexpr bool homogeneousCFBC = false;
 
-  // Update ghost cells and jump conditions first. Doing an exchange is not sufficient here because
-  // the jump stencils might reach over CFs.
   this->exchangeGhost(a_phi);
-  this->interpolateCF(a_phi, &a_phiCoar, homogeneousCFBC);
   this->updateJumpBC(a_phi, a_homogeneousPhysBC);
 
   for (auto& op : m_helmOps) {
@@ -1091,8 +1086,7 @@ MFHelmholtzOp::AMROperatorNC(LevelData<MFCellFAB>&             a_Lphi,
[…]
