### 2023-05-10 d9b770e
Improved coarsening to ebhelmholtzopfactory
