### 2023-05-10 128e4c1
Added SOR factor to EBHelmholtzOp. Default is 1. Will not be used by anyone, I think
