+++ Source/AmrMesh/CD_EBCoarAve.cpp
@@ -688,8 +688,8 @@ EBCoarAve::arithmeticAverage(EBFaceFAB&       a_coarData,
 
   // Kernel for doing regular faces.
   const Real invFinePerCoar = 1.0 / std::pow(m_refRat, SpaceDim - 1);
-  const int xDoLoop        = (a_dir == 0) ? 0 : 1;
-  const int yDoLoop        = (a_dir == 1) ? 0 : 1;
+  const int  xDoLoop        = (a_dir == 0) ? 0 : 1;
+  const int  yDoLoop        = (a_dir == 1) ? 0 : 1;
 #if CH_SPACEDIM == 3
   const int zDoLoop = (a_dir == 2) ? 0 : 1;
 #endif
@@ -840,8 +840,8 @@ EBCoarAve::conservativeAverage(EBFaceFAB&       a_coarData,
 
   // Kernel for doing regular faces.
   const Real invFinePerCoar = 1.0 / std::pow(m_refRat, SpaceDim - 1);
-  const int xDoLoop        = (a_dir == 0) ? 0 : 1;
-  const int yDoLoop        = (a_dir == 1) ? 0 : 1;
+  const int  xDoLoop        = (a_dir == 0) ? 0 : 1;
+  const int  yDoLoop        = (a_dir == 1) ? 0 : 1;
 #if CH_SPACEDIM == 3
   const int zDoLoop = (a_dir == 2) ? 0 : 1;
 #endif
