### 2023-05-15 b26a703
changed computeDt to different computation before and after Vpeak. Updated m_prevDt to initialisation in the constructor
