+++ Source/AmrMesh/CD_EBMGRestrict.H
@@ -96,11 +96,6 @@ protected:
   EBLevelGrid m_eblgCoFi;
 
   /*!
-    @brief Buffer data on the coarsened fine grids
-  */
-  mutable LevelData<EBCellFAB> m_dataCoFi;
-
-  /*!
     @brief Iterator for iterating over cut-cells on the coarse level
   */
   mutable LayoutData<VoFIterator> m_vofitCoar;
