### 2023-05-19 #5335 @rmrsk
@ppxasjsm Here is the extra information:

Release version: v23.05
DOI: [10.5281/zenodo.7950220](https://doi.org/10.5281/zenodo.7950220)

Regarding license: Could not find GPL-3.0 in the dropdown list on Zenodo but the guidelines says to choose "other", and include the LICENSE file with the archived files. Which I've done.
