+++ README.md
@@ -4,6 +4,8 @@ chombo-discharge
 This is ``chombo-discharge``, a multiphysics code which uses ``Chombo`` for discharge simulations with adaptive mesh refinement (AMR) on embedded boundary grids.
 A modified version of ``Chombo`` is distributed together with this code.
 
+[![DOI](https://zenodo.org/badge/355565674.svg)](https://zenodo.org/badge/latestdoi/355565674)
+
 <p align="center">
    <img src="./Docs/Sphinx/source/_static/figures/BranchingAir.gif" width="50%">
 </p align="center">
