### 2023-05-20 #5335 @ppxasjsm
Yeah I don't think it contains a GPL-3 option. I have used GPL-2 or later as an option before, but including your file should be fine.

 Thanks for creating the release version and archive!
