### 2023-05-21 #5335 @rmrsk
@ppxasjsm My bad - did not catch the "... or later" part of it. Changed it to GPL-2.0 or later just now. Decided to leave FannySkirbekk out since the contribution was comparatively minor.
