### 2023-05-22 ab619c6
Update paper.md
Minor readability / typographic edits to JOSS paper
