+++ paper/paper.bib
@@ -64,6 +64,8 @@ abstract = {A 3D fluid model has been developed to simulate streamer discharges 
 }
 @misc{ebchombo,
    author = {{Colella}, P and {Graves}, D T and {Ligocki}, T J and {Miller}, G H and {Modiano}, D and {Schwartz}, P O and {Van Straalen}, B and {Pillod}, J and {Trebotich}, D and {Barad}, M},
-   title = {EBChombo Software Package for Cartesian Grid, Embedded Boundary Applications (LBNL-6615E)},   
-   publisher = {Lawrence Berkelely National Laboratory},
+   title = {EBChombo Software Package for Cartesian Grid, Embedded Boundary Applications},   
+   publisher = {Lawrence Berkeley National Laboratory},
+   howpublished = {Technical Report LBNL-6615E},
+   year = 2000
 }
