### 2023-05-22 #5335 @kyleniemeyer
Hi @rmrsk, I made some minor edits to the paper in https://github.com/chombo-discharge/chombo-discharge/pull/332

Can you review and merge this PR? We can then finish the publication process.
