### 2023-05-23 #5335 @kyleniemeyer
@rmrsk ok, it looks like the paper PDF has been updated (https://www.theoj.org/joss-papers/joss.05335/10.21105.joss.05335.pdf); the cached version at https://doi.org/10.21105/joss.05335 might take a bit longer to reflect the correction.
